% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotate_head}
\alias{annotate_head}
\title{Add a head shape to a ggplot}
\usage{
annotate_head(size = 1.1, color = "black", stroke = 1)
}
\arguments{
\item{size}{Size of the head}

\item{color}{Color of the head}

\item{stroke}{Line thickness}
}
\value{
A layer for a ggplot
}
\description{
Adds the outline of a head and nose to a ggplot.
}
\examples{
library(ggplot2)
data_faces_ERPs \%>\%
  eeg_filter(between(as_time(.sample, .unit = "milliseconds"), 100, 200)) \%>\%
  eeg_group_by(condition) \%>\%
  eeg_summarize(across_ch(mean, na.rm = TRUE)) \%>\%
  plot_topo() +
  annotate_head(size = .9, color = "black", stroke = 1) +
  annotate_electrodes()
}
\seealso{
Other plotting functions:
\code{\link[=annotate_electrodes]{annotate_electrodes()}},
\code{\link[=annotate_events]{annotate_events()}},
\code{\link[=eeg_browse]{eeg_browse()}},
\code{\link[=eeg_downsample]{eeg_downsample()}},
\code{\link[=ggplot.eeg_lst]{ggplot.eeg_lst()}},
\code{\link[=plot.eeg_lst]{plot.eeg_lst()}},
\code{\link[=plot_components]{plot_components()}},
\code{\link[=plot_in_layout]{plot_in_layout()}},
\code{\link[=plot_topo]{plot_topo()}},
\code{\link[=theme_eeguana]{theme_eeguana()}}
}
\concept{plotting functions}
