% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trans.R
\name{as.data.frame.eegbl}
\alias{as.data.frame.eegbl}
\title{Convert an eegble to a (base) data frame.}
\usage{
\method{as.data.frame}{eegbl}(x, ..., thinning = "auto",
  add_seg_info = TRUE)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{...}{Other arguments passed on to individual methods.

`as_data_frame` and `as.tibble` are aliases.}

\item{thinning}{}

\item{add_seg_info}{}
}
\value{
A tibble.
}
\description{
Convert an eegble to a (base) data frame.
}
