% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as.data.table.eeg_lst}
\alias{as.data.table.eeg_lst}
\title{Convert an eeg_lst to a long table in \code{\link[data.table:data.table]{data.table}} format.}
\usage{
as.data.table.eeg_lst(x, .unit = "s", ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{.unit}{Unit for the \code{.time} column of the transformed object: "s" (default), "ms", "samples".}

\item{...}{Additional arguments to be passed to or from other methods.}
}
\value{
A \code{\link[data.table:data.table]{data.table}}.
}
\description{
Convert the signal_tbl table from wide to long format.
}
