% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{as_eeg_lst}
\alias{as_eeg_lst}
\title{Transforms an object to a eeg_lst}
\usage{
as_eeg_lst(.data, ...)
}
\arguments{
\item{.data}{An \code{eeg_ica_lst} and experimentally an MNE raw signal using \link[reticulate:reticulate]{reticulate::reticulate}.}

\item{...}{Not in use.}
}
\description{
Transforms an object to a eeg_lst
}
