% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_sample.R
\name{as_sample_int}
\alias{as_sample_int}
\alias{as_sample_int.numeric}
\title{Convert a time point  into a sample.}
\usage{
as_sample_int(x, ...)

\method{as_sample_int}{numeric}(x, sampling_rate = NULL, unit = "s", ...)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{...}{Not in use.}

\item{sampling_rate}{Sampling rate in Hz}

\item{unit}{"seconds" (or "s"), "milliseconds" (or "ms"), or "samples"}
}
\value{
A sample_int object.
}
\description{
Convert a time point  into a sample.
}
