% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as_tibble.eeg_lst}
\alias{as_tibble.eeg_lst}
\alias{as_data_frame.eeg_lst}
\alias{as_long_tbl.eeg_lst}
\title{Convert an eeg_lst to a long table in \code{\link[tibble:tibble]{tibble}} format.}
\usage{
as_tibble.eeg_lst(x, .unit = "s", ...)

as_data_frame.eeg_lst(x, .unit = "s", ...)

\method{as_long_tbl}{eeg_lst}(x, .unit = "s", ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{.unit}{Unit for the \code{.time} column of the transformed object: "s" (default), "ms", "samples".}

\item{...}{Additional arguments to be passed to or from other methods.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}
}
\description{
Convert the signal_tbl table from wide to long format.
}
\seealso{
Other tibble:
\code{\link[=as.data.frame.psd_lst]{as.data.frame.psd_lst()}},
\code{\link[=as_tibble.psd_lst]{as_tibble.psd_lst()}}
}
\concept{tibble}
