% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trans.R
\name{as_tibble.eegbl}
\alias{as_tibble.eegbl}
\alias{as_data_frame.eegbl}
\alias{as.tibble.eegbl}
\title{Convert an eegble to a tibble.}
\usage{
\method{as_tibble}{eegbl}(x, ..., thinning = NULL, add_seg_info = TRUE)

as_data_frame.eegbl(x, ..., thinning = NULL, add_seg_info = TRUE)

as.tibble.eegbl(x, ..., thinning = NULL, add_seg_info = TRUE)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{...}{Other arguments passed on to individual methods.

`as_data_frame` and `as.tibble` are aliases.}

\item{thinning}{}

\item{add_seg_info}{}
}
\value{
A tibble.
}
\description{
Convert an eegble to a tibble.
}
