% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as_tibble.psd_lst}
\alias{as_tibble.psd_lst}
\title{Convert an psd_lst to a long table in \code{\link[tibble:tibble]{tibble}} format.}
\usage{
as_tibble.psd_lst(x, ...)
}
\arguments{
\item{x}{A \code{psd_lst} object.}

\item{...}{Additional arguments to be passed to or from other methods.}
}
\value{
A \code{\link[tibble:tibble]{tibble}}
}
\description{
Convert the signal_tbl table from wide to long format.
}
\seealso{
Other tibble:
\code{\link[=as.data.frame.psd_lst]{as.data.frame.psd_lst()}},
\code{\link[=as_tibble.eeg_lst]{as_tibble.eeg_lst()}}
}
\concept{tibble}
