% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as_tidytable.eeg_lst}
\alias{as_tidytable.eeg_lst}
\title{Convert an eeg_lst to a long table in \code{\link[tidytable:tidytable]{tidytable}} format.}
\usage{
as_tidytable.eeg_lst(x, .unit = "s")
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{.unit}{Unit for the \code{.time} column of the transformed object: "s" (default), "ms", "samples".}
}
\value{
A \code{\link[tidytable:tidytable]{tidytable}}.
}
\description{
Convert the signal_tbl table from wide to long format.
}
