% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as_tidytable.psd_lst}
\alias{as_tidytable.psd_lst}
\title{Convert an psd_lst to a long table in \code{\link[tidytable:tidytable]{tidytable}} format.}
\usage{
as_tidytable.psd_lst(x, ...)
}
\arguments{
\item{x}{A \code{psd_lst} object.}
}
\value{
A \code{\link[tidytable:tidytable]{tidytable}}.
}
\description{
Convert the signal_tbl table from wide to long format.
}
