% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_sample.R
\name{as_time}
\alias{as_time}
\title{Convert a sample point into a time point.}
\usage{
as_time(x, .unit = "second")
}
\arguments{
\item{x}{A \code{sample_id} object.}

\item{.unit}{"seconds" (or "s"), "milliseconds" (or "ms")}
}
\value{
A vector of times.
}
\description{
Convert a sample point into a time point.
}
