% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_trans.R
\name{baseline}
\alias{baseline}
\title{Baseline an eegble.}
\usage{
baseline(x, t = -Inf, parallel = TRUE)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{t}{A negative number indicating from when to baseline 
(TO COMPLETE). The default is to use all the negative times.}

\item{parallel}{If set to TRUE, it will use all cores. 
(Default: parallel = TRUE).}
}
\value{
An eegbl.
}
\description{
Baseline an eegble.
}
