% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{between}
\alias{between}
\title{Convenience function for range subsets}
\usage{
between(x, lower, upper, incbounds = TRUE)
}
\arguments{
\item{x}{ Any orderable vector, i.e., those with relevant methods for
\code{`<=`}, such as \code{numeric}, \code{character}, \code{Date}, etc. in
case of \code{between} and a numeric vector in case of \code{inrange}.}

\item{lower}{ Lower range bound.}

\item{upper}{ Upper range bound.}

\item{incbounds}{ \code{TRUE} means inclusive bounds, i.e., [lower,upper].
\code{FALSE} means exclusive bounds, i.e., (lower,upper).

It is set to \code{TRUE} by default for infix notations.}
}
\description{
between is a thin wrapper for the between function of \link{data.table}. It is equivalent to x >= lower & x <= upper when incbounds=TRUE, or x > lower & y < upper when FALSE.
}
