% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_ext.R
\name{bind}
\alias{bind}
\title{Bind eeg_lst objects.}
\usage{
bind(...)
}
\arguments{
\item{...}{\code{eeg_lst} objects to combine.}
}
\value{
An \code{eeg_lst} object.
}
\description{
Binds eeg_lst and throws a warning if there is a mismatch in the channel information.
}
