% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_transformation.R
\name{ch_baseline}
\alias{ch_baseline}
\alias{ch_baseline.eeg_lst}
\alias{ch_baseline.channel_dbl}
\title{Baseline an eeg_lst}
\usage{
ch_baseline(x, ...)

\method{ch_baseline}{eeg_lst}(x, time = -Inf, sample_id = NULL, ...)

\method{ch_baseline}{channel_dbl}(x, time = -Inf, sample_id = NULL,
  ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object or a channel.}

\item{...}{Not in use.}

\item{time}{A negative number indicating from when to baseline; the interval is defined as [time,0]. The default is to use all the negative times.}

\item{sample_id}{A negative number indicating from when to baseline. The default is to use all the negative times. (\code{time} is ignored if  sample_id is used).}
}
\value{
An eeg_lst.
}
\description{
Subtract the average or baseline of the points in a defined interval from all points in the segment.
}
\seealso{
Other channel: \code{\link{channel_dbl}},
  \code{\link{chs_mean}}, \code{\link{eeg_rereference}},
  \code{\link{is_channel_dbl}}
}
\concept{channel}
