% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_processing.R
\name{ch_rereference}
\alias{ch_rereference}
\title{Rereference channel.}
\usage{
ch_rereference(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{A channel.}

\item{...}{Channels that will be averaged as the reference.}
}
\value{
A rereferenced channel.
}
\description{
This function is meant to be used together with \code{mutate} or \code{mutate_all}. See the example
}
\examples{
\dontrun{
# Rereference all channels used the linked mastoids (average of the two mastoids)

faces_segs \%>\% act_on(signal_tbl) \%>\%
                 mutate_all(funs(rereference(., M1, M2)))
}#'
}
\seealso{
Other channel: \code{\link{ch_baseline}},
  \code{\link{channel_dbl}}, \code{\link{chs_mean}},
  \code{\link{is_channel_dbl}}
}
\concept{channel}
