% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{change_coord}
\alias{change_coord}
\title{Change coordinate system from 3D to 2D}
\usage{
change_coord(data, .projection = "polar")
}
\arguments{
\item{data}{A tbl created with .keys_tbl}

\item{.projection}{projection type}
}
\value{
A modified .keys_tbl
}
\description{
Change coordinate system from 3D to 2D
}
