% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{chs_mean}
\alias{chs_mean}
\title{Get the by-sample (or by-row) mean of the specified channels.}
\usage{
chs_mean(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{...}{A group of channels, it can be used in combination with \code{across}.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE}
    indicating whether \code{NA} values should be stripped before the
    computation proceeds.}
}
\value{
A new channel or an \code{eeg_lst} object with a \code{mean} channel instead of the previous channels.
}
\description{
Wrapper of \code{rowMeans} that performs a by-sample mean of the specified channels.
}
\examples{
\dontrun{
faces_segs_some \%>\%
  eeg_transmute(
    Occipital = chs_mean(O1, O2, Oz, na.rm = TRUE),
    Parietal = chs_mean(P3, P4, P7, P8, Pz, na.rm = TRUE)
  )

faces_segs_some \%>\%
  eeg_transmute(
    Occipital = chs_mean(O1, O2, Oz, na.rm = TRUE),
    Parietal = chs_mean(P3, P4, P7, P8, Pz, na.rm = TRUE)
  )
faces_seg \%>\%
  eeg_transmute(
    Occipital = chs_mean(across(starts_with("O")), na.rm = TRUE), # O1, O2, Oz
    Parietal = chs_mean(across(starts_with("O")), na.rm = TRUE) # P3, P4, P7, P8, Pz
  )
faces_segs_some \%>\%
  chs_mean(na.rm = TRUE)
}
}
\seealso{
Other channel functions:
\code{\link[=chs_fun]{chs_fun()}}
}
\concept{channel functions}
