% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_faces_ERPs}
\alias{data_faces_ERPs}
\title{ERPs elicited when one subject saw faces and non-faces objects.}
\format{
An eeg_lst  with 2 conditions (faces, non-faces)
}
\usage{
data(data_faces_ERPs)
}
\description{
A dataset containing the ERPs elicited when one subject saw faces and non-faces objects. See also the vignette \href{https://bnicenboim.github.io/eeguana/articles/brainvision_files.html}{Introduction - Manipulating a clean EEG file exported from BrainVision 2.0}.
}
\seealso{
Other EEG datasets:
\code{\link{data_faces_10_trials}}
}
\concept{EEG datasets}
\keyword{datasets}
