% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{dplyr_groups}
\alias{dplyr_groups}
\alias{groups.eeg_lst}
\alias{group_vars.eeg_lst}
\title{Return grouping variables.}
\usage{
groups.eeg_lst(x)

group_vars.eeg_lst(x)
}
\arguments{
\item{x}{An eeg_lst.}
}
\description{
Return grouping variables.
}
