% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{eeg_lst}
\alias{eeg_lst}
\title{Create an eeg_lst}
\usage{
eeg_lst(
  signal_tbl = NULL,
  events_tbl = NULL,
  segments_tbl = NULL,
  channels_tbl = NULL,
  .sampling_rate = NULL
)
}
\arguments{
\item{signal_tbl}{See \code{\link[=signal_tbl]{signal_tbl()}}.}

\item{events_tbl}{See \code{\link[=events_tbl]{events_tbl()}}.}

\item{segments_tbl}{A data table of segment numbers and related information. See \code{\link[=segments_tbl]{segments_tbl()}}.}

\item{channels_tbl}{Optionally a table with channels information. See \code{\link[=channels_tbl]{channels_tbl()}}.}

\item{.sampling_rate}{Optional: If the signal_tbl doesn't have samples, they will be included with this sampling rate.}
}
\value{
A valid eeg_lst.
}
\description{
Builds an eeg_lst object composed of two \code{data.table::data.table} objects and one  \code{tibble::tibble}. All three are linked by a unique identifier \code{.id}. Amplitude values and timestamps appear in the \code{signal} table. Triggers, blinks, artifact rejection markings, and other events logged by the EEG recording software appear in the \code{events} table. Segment information and recording IDs appear in the \code{segments} tibble.
}
\details{
The \code{signal} table is organized into columns representing timestamps (\code{.sample}) and individual electrodes. Each \code{.sample} corresponds to 1 sample in the original recording, i.e. if the sampling rate of the EEG recording is 500 Hz, then each \code{.sample} corresponds to 2 milliseconds. These timestamps correspond to \code{.initial} in the \code{events} table, which displays only the timestamps where logged events began.

The \code{events} table is organized into columns representing the \code{.type} of event associated with the trigger listed under \code{.description}. The timestamp marking the beginning and the end of the event is listed under \code{.initial} and \code{.final} (in samples). The \code{.channel} column is a  linking variable only, so will generally only contain NAs, unless the event is specific to a certain channel.

The \code{segments} tibble contains the subject ID under \code{recording}, which is the file name unless otherwise specified. If the data has been segmented in BrainVision, the segment number will be listed under \code{segment}. The data can also be segmented according to trigger labels in \code{eeguana}, see \code{segment}. \code{segment} will be place the segment number under \code{segment}, the trigger name under \code{.type.x}, and the trigger label under \code{.description.x}. Other information such as condition labels or response times can be added by the user by merging into the \code{segments} tibble using non-eeguana merge functions, e.g. the \code{dplyr} join series.
}
\seealso{
Other eeg_lst:
\code{\link[=is_eeg_lst]{is_eeg_lst()}}
}
\concept{eeg_lst}
