% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeguana.R
\docType{package}
\name{eeguana}
\alias{eeguana}
\alias{eeguana-package}
\title{eeguana: A package for flexible manipulation of EEG data.}
\description{
The eeguana package provides functions that can be used on already pre-processed (i.e.,
filtering, artifact rejects, ICA, etc has already been done) EEG files (at least for now).
The package mainly provides dplyr-like functions to manipulate the EEG data, and ggplot wrapper functions.
}
