% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_trans.R
\name{event_to_NA}
\alias{event_to_NA}
\title{Remove (transform to NA) problematic events from an eegble.}
\usage{
event_to_NA(x, ..., ind_channel = TRUE, entire_seg = FALSE,
  parallel = FALSE)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{...}{Description of the problematic event.}

\item{ind_channel}{If set to FALSE, 
it will remove samples from all channels (Default:  ind_channel = TRUE).}

\item{entire_seg}{If set to TRUE, it will remove the entire segment 
(Default: entire_seg = FALSE).}

\item{parallel}{If set to TRUE, it will use all cores. 
   (Default: parallel = FALSE). See details.

The parallel option runs the process in several cores using \code{furrr} 
with \code{future} as the backends. To use the parallel option these 
packages need to be installed, the library \code{future} needs to be 
attached, and a "plan" needs to be specified (e.g., 
\code{plan(multiprocess)})}
}
\value{
An eegbl.
}
\description{
Remove (transform to NA) problematic events from an eegble.
}
\examples{
\dontrun{
library(future)
plan(multiprocess)
clean_eggble <- event_to_NA(eegble, parallel = TRUE)
}
}
