% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{events}
\alias{events}
\alias{events.eeg_lst}
\alias{events<-}
\alias{events<-.eeg_lst}
\title{Functions to get or set the events table of an eeg_lst object.}
\usage{
events(x, ...)

\method{events}{eeg_lst}(x)

events(x, ...) <- value

\method{events}{eeg_lst}(x) <- value
}
\arguments{
\item{x}{An eeg_lst object.}

\item{value}{An events table.}
}
\value{
A table.
}
\description{
Functions to get or set the events table of an eeg_lst object.
}
