% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{ggplot.eeg_lst}
\alias{ggplot.eeg_lst}
\title{Create an ERP plot}
\usage{
\method{ggplot}{eeg_lst}(data = NULL, mapping = ggplot2::aes(), ..., .max_sample = 64000)
}
\arguments{
\item{data}{An \code{eeg_lst} object.}

\item{mapping}{Default list of aesthetic mappings to use for plot.
If not specified, must be supplied in each layer added to the plot.}

\item{...}{Other arguments passed on to methods. Not currently used.}

\item{.max_sample}{Downsample to approximately 6400 samples by default.}
}
\value{
A ggplot object
}
\description{
\code{ggplot} initializes a ggplot object which takes an \code{eeg_lst} object as
its input data. Layers can then be added in the same way as for a
\link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\details{
If necessary, t will first downsample the \code{eeg_lst} object so that there is a
maximum of 6400 samples. The \code{eeg_lst} object is then converted to a long-format
tibble via \link[tibble:as_tibble]{as_tibble}. In this tibble, the \code{.key} variable is the
channel/component name and \code{.value} its respective amplitude. The sample
number (\code{.sample} in the \code{eeg_lst} object) is automatically converted to milliseconds
to create the variable \code{.time}. By default, time is plotted on the
x-axis and amplitude on the y-axis.

To add additional components to the plot such as titles and annotations, simply
use the \code{+} symbol and add layers exactly as you would for \link[ggplot2:ggplot]{ggplot2::ggplot}.
}
\examples{
library(ggplot2)
# Plot grand averages for selected channels
data_faces_ERPs \%>\%
  # select the desired electrodes
  eeg_select(O1, O2, P7, P8) \%>\%
  ggplot(aes(x = .time, y = .key)) +
  # add a grand average wave
  stat_summary(
    fun.y = "mean", geom = "line", alpha = 1, linewidth = 1.5,
    aes(color = condition)
  ) +
  # facet by channel
  facet_wrap(~.key) +
  theme(legend.position = "bottom")
  
}
\seealso{
Other plotting functions:
\code{\link[=annotate_electrodes]{annotate_electrodes()}},
\code{\link[=annotate_events]{annotate_events()}},
\code{\link[=annotate_head]{annotate_head()}},
\code{\link[=eeg_browse]{eeg_browse()}},
\code{\link[=eeg_downsample]{eeg_downsample()}},
\code{\link[=plot.eeg_lst]{plot.eeg_lst()}},
\code{\link[=plot_components]{plot_components()}},
\code{\link[=plot_in_layout]{plot_in_layout()}},
\code{\link[=plot_topo]{plot_topo()}},
\code{\link[=theme_eeguana]{theme_eeguana()}}
}
\concept{plotting functions}
