% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{ica_matrix_lst}
\alias{ica_matrix_lst}
\title{Function to get the list of mixing and unmixing matrices of an eeg_ica_lst object.}
\usage{
ica_matrix_lst(.data, ...)
}
\arguments{
\item{.data}{An eeg_ica_lst object.}

\item{...}{Not in use.}
}
\value{
A list.
}
\description{
Function to get the list of mixing and unmixing matrices of an eeg_ica_lst object.
}
\seealso{
Other functions to expose internal parts of eeg_(ica_)_lst:
\code{\link[=channels_tbl]{channels_tbl()}},
\code{\link[=events_tbl]{events_tbl()}},
\code{\link[=segments_tbl]{segments_tbl()}},
\code{\link[=signal_tbl]{signal_tbl()}}

Other ica methods:
\code{\link{fICA}},
\code{\link[=fast_ICA]{fast_ICA()}}
}
\concept{functions to expose internal parts of eeg_(ica_)_lst}
\concept{ica methods}
