% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{info}
\alias{info}
\alias{nchannels}
\alias{channel_names}
\alias{channels}
\alias{info}
\alias{srate}
\alias{duration}
\title{Display information of the eegble object.}
\usage{
nchannels(x, ...)

channel_names(x, ...)

channels(x, ...)

info(x, ...)

srate(x, ...)

duration(x, ...)
}
\arguments{
\item{x}{An eegble object.}
}
\description{
\itemize{
\item \code{nchannels()}: Returns the number of channels.
\item \code{chan_names()}: Returns a vector with the name of the channels.
\item \code{channels()}: Returns a data frame (tibble) with information about the channels.
\item \code{info()}: Returns a data frame (tibble) with information about the EEG recording.
\item \code{srate()}: Returns the sampling rate.
\item \code{duration()}: Returns the duration of the recording (or segments).
}
}
