% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{install_py_eeguana}
\alias{install_py_eeguana}
\title{Install the Python packages needed for \code{eeguana}}
\usage{
install_py_eeguana(
  conda = "auto",
  envname = "r-eeguana",
  restart_session = TRUE,
  forge = TRUE,
  ...,
  new_env = identical(envname, "r-eeguana"),
  python_version = NULL
)
}
\arguments{
\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda_binary]{conda_binary()}} for more details.}

\item{envname}{The name of the virtual environment to create or use. Default is 'r-eeguana'.}

\item{restart_session}{Whether to restart the R session after installation. Default is TRUE.}

\item{forge}{Whether to use conda-forge to install packages. Default is TRUE.}

\item{...}{Additional arguments passed to \code{reticulate::py_install}.}

\item{new_env}{Whether to create a new environment if \code{envname} is 'r-eeguana'. Default behavior is determined by the identity of \code{envname}.}

\item{python_version}{The requested Python version. Ignored when attempting
to install with a Python virtual environment.}
}
\value{
Returns \code{NULL} invisibly and prints a message upon successful installation. If \code{restart_session} is TRUE and the R session is running within RStudio, the session will restart automatically.
}
\description{
\code{install_py_eeguana} facilitates the installation of Python packages required by \code{eeguana} within an R environment.
It leverages the \code{reticulate} package to manage Python environments and supports various installation methods,
environment configurations, and Python versions.
}
\details{
The function selects an appropriate method for environment management and Python installation, focusing on ease of use and flexibility.
It supports creating new virtual or conda environments, installing necessary packages, and handling Python version requirements.

The function is designed to be robust, handling various scenarios such as existing environments, and provides detailed feedback during the process.
}
