% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{is_component_dbl}
\alias{is_component_dbl}
\title{Test if the object is a component
This function returns  TRUE for components.}
\usage{
is_component_dbl(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{sample_id} class.
}
\description{
Test if the object is a component
This function returns  TRUE for components.
}
\seealso{
Other component:
\code{\link[=component_dbl]{component_dbl()}}
}
\concept{component}
