% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{is_eeg_lst}
\alias{is_eeg_lst}
\title{Test if the object is an eeg_lst.}
\usage{
is_eeg_lst(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{eeg_lst} class.
}
\description{
This function returns  TRUE for eeg_lsts.
}
\seealso{
Other eeg_lst:
\code{\link[=eeg_lst]{eeg_lst()}}
}
\concept{eeg_lst}
