% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/events_tbl.R
\name{is_events_tbl}
\alias{is_events_tbl}
\title{Test if the object is an events_tbl
This function returns  TRUE for events_tbl.}
\usage{
is_events_tbl(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{events_tbl} class.
}
\description{
Test if the object is an events_tbl
This function returns  TRUE for events_tbl.
}
\concept{events_tbl}
