% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{is_psd_tbl}
\alias{is_psd_tbl}
\title{Test if the object is a  psd_tbl}
\usage{
is_psd_tbl(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{psd_tbl} class.
}
\description{
This function returns  TRUE for psds.
}
\concept{psd_tbl}
