% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{join-eeguana}
\alias{join-eeguana}
\alias{anti_join.eeg_lst}
\alias{left_join.eeg_lst}
\alias{semi_join.eeg_lst}
\title{Dplyr functions for joining data frames to the segments of  eeg_lst objects.}
\usage{
anti_join.eeg_lst(x, y, by = NULL, copy = FALSE, suffix = c(".x",
  ".y"), ...)

left_join.eeg_lst(x, y, by = NULL, copy = FALSE, suffix = c(".x",
  ".y"), ...)

semi_join.eeg_lst(x, y, by = NULL, suffix = c(".x", ".y"), ...)
}
\arguments{
\item{x}{An eeg_lst.}

\item{y}{A data frame, tibble, or data.table.}

\item{by}{a character vector of variables to join by.  If \code{NULL}, the
default, \code{*_join()} will do a natural join, using all variables with
common names across the two tables. A message lists the variables so
that you can check they're right (to suppress the message, simply
explicitly list the variables that you want to join).

To join by different variables on x and y use a named vector.
For example, \code{by = c("a" = "b")} will match \code{x.a} to
\code{y.b}.}

\item{copy}{If \code{x} and \code{y} are not from the same data source,
and \code{copy} is \code{TRUE}, then \code{y} will be copied into the
same src as \code{x}.  This allows you to join tables across srcs, but
it is a potentially expensive operation so you must opt into it.}

\item{suffix}{If there are non-joined duplicate variables in \code{x} and
\code{y}, these suffixes will be added to the output to disambiguate them.
Should be a character vector of length 2.}

\item{...}{other parameters passed onto methods, for instance, \code{na_matches}
to control how \code{NA} values are matched.  See \link{join.tbl_df} for more.}
}
\value{
An eeg_lst object.
}
\description{
Join a data frames to the segments of an eeg_lst object, and
modify the signal table accordingly (dropping rows when necessary).
}
\details{
Wrappers for \link[dplyr:join]{dplyr}'s commands:
The following wrappers have been implemented for \code{eeg_lst} objects:
\itemize{
\item \code{left_join()}: left-joins an external table to the segments table of the eeg_lst.
\item \code{semi_join()}: semi-joins an external table to the segments table of the eeg_lst.
\item \code{anti_join()}: anti-joins an external table to the segments table of the eeg_lst.
These commands always return the entire eeg_lst so that
they can be ' piped using \link[magrittr:magrittr]{magrittr} 's pipe, %>%.
}
}
\seealso{
Other dplyr functions: \code{\link{dplyr_verbs}}
}
\concept{dplyr functions}
