% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{layout_32_1020}
\alias{layout_32_1020}
\title{Layout for a 32 electrodes cap at the standard 10-20 system}
\format{
An table with channel names and locations
}
\usage{
data(layout_32_1020)
}
\description{
A dataset containing coordinates of a 32 electrodes mounted in an elastic cap at the standard 10-20 system (Jasper, 1958)
}
\seealso{
Other topographic plots and layouts:
\code{\link[=plot_components]{plot_components()}},
\code{\link[=plot_in_layout]{plot_in_layout()}},
\code{\link[=plot_topo]{plot_topo()}}
}
\concept{topographic plots and layouts}
\keyword{datasets}
