% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{mixing_tbl}
\alias{mixing_tbl}
\title{Builds a mixing_tbl table.}
\usage{
mixing_tbl(mixing_matrix, means_matrix, groups, channels_tbl)
}
\arguments{
\item{mixing_matrix}{Matrix or table of channels with their mixing.}

\item{ids}{Integers indicating to which group the row of the mixing matrix belongs.}

\item{sample_ids}{Vector of integers.}

\item{channel_info}{A table with information about each channel (such as the one produced by `channels_tbl``)}
}
\value{
A valid mixing_tbl table.
}
\description{
Builds a mixing_tbl table.
}
\concept{mixing_tbl}
