% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_wrangling.R
\name{mutate.eegbl}
\alias{mutate.eegbl}
\title{Mutate the signal of an eegble.}
\usage{
\method{mutate}{eegbl}(.data, ...)
}
\arguments{
\item{.data}{An eegbl.}

\item{...}{Name-value pairs of expressions, tipically they will include channel names.}
}
\value{
An eegbl.
}
\description{
Mutate the signal of an eegble.
}
