% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_wrangling.R
\name{mutate_chan}
\alias{mutate_chan}
\title{Mutate the signal of an eegble.}
\usage{
mutate_chan(.data, ...)
}
\arguments{
\item{.data}{An eegbl.}

\item{...}{Name-value pairs of expressions, tipically they will include channel names.}
}
\value{
An eegbl.
}
\description{
Mutate the signal of an eegble.
}
