% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{na_omit}
\alias{na_omit}
\title{Handle Missing Values in eeg_lst objects}
\usage{
na_omit(object, ...)
}
\arguments{
\item{object}{an \code{eeg_lst}.}

\item{...}{further arguments special methods could require.}
}
\description{
Remove all NAs from an \code{eeg_lst} signal table.
}
