% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{plot.eegbl}
\alias{plot.eegbl}
\title{Simple plot an eegble object.}
\usage{
\method{plot}{eegbl}(x, thinning = "auto")
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{thinning}{Automatic by default, but integers can be used.}
}
\value{
A ggplot object
}
\description{
Simple plot an eegble object.
}
