% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gg}
\alias{plot_gg}
\alias{plot_gg.eeg_lst}
\title{Create an ERP plot}
\usage{
plot_gg(.data, ...)

\method{plot_gg}{eeg_lst}(.data, x = time, y = .value, ...,
  max_sample = 64000)
}
\arguments{
\item{.data}{An \code{eeg_lst} object.}

\item{...}{List of name value pairs giving aesthetics to map to
variables. The names for x and y aesthetics are typically omitted because
they are so common; all other aesthetics must be named.}

\item{x}{List of name value pairs giving aesthetics to map to
variables. The names for x and y aesthetics are typically omitted because
they are so common; all other aesthetics must be named.}

\item{y}{List of name value pairs giving aesthetics to map to
variables. The names for x and y aesthetics are typically omitted because
they are so common; all other aesthetics must be named.}

\item{max_sample}{Downsample to approximately 64000 samples by default.}
}
\value{
A ggplot object
}
\description{
\code{plot_gg} initializes a ggplot object which takes an eeg_lst object as
its input data. Layers can then be added in the same way as for a ggplot2
\code{\link[ggplot2]{ggplot}} object.
}
\examples{
# Download and segment data as per 
https://bnicenboim.github.io/eeguana/articles/brainvision_files.html

# Plot grand averages over raw data in selected channels
faces_segs_some \%>\% 
  select(O1, O2, P7, P8) \%>\% 
  plot_gg() + 
      geom_line(alpha = .1, aes(group = .id, color = condition)) + 
      stat_summary(fun.y = "mean", geom ="line", alpha = 1, size = 1.5, 
               aes(color = condition)) +
      facet_wrap(~ channel) + 
      geom_vline(xintercept = 0, linetype = "dashed") + 
      geom_vline(xintercept = .17, linetype = "dotted") + 
      theme(legend.position = "bottom") 

}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_in_layout}}, \code{\link{plot_topo}}
}
\concept{plot}
