% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{print.eegbl}
\alias{print.eegbl}
\title{Print an eegble object.}
\usage{
\method{print}{eegbl}(x, ..., file = NULL)
}
\arguments{
\item{x}{An eegble object.}

\item{...}{Other options passed to print.tbl.}

\item{file}{If it is specified, prints a summary of a single file.}
}
\description{
Print an eegble object.
}
