% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_helpers.R
\name{read_dat}
\alias{read_dat}
\title{Helper function to read the dat files directly}
\usage{
read_dat(file, header_info = NULL, events = NULL, recording, sep, zero)
}
\description{
Helper function to read the dat files directly
}
