% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_vhdr}
\alias{read_vhdr}
\title{Read a BrainVision file into R}
\usage{
read_vhdr(
  file,
  .sep = .type == "New Segment",
  .zero = .type == "Time 0",
  .recording = file
)
}
\arguments{
\item{file}{A vhdr file in a folder that contains a .vmrk and .dat files}

\item{.sep}{Segment separation marker. By default: \code{.type == "New Segment"}}

\item{.zero}{Time zero marker. By default: \code{.type == "Time 0"}}

\item{.recording}{Recording name (\code{file} name, by default).}
}
\value{
An \code{eeg_lst} object with signal_tbl and event from file_name.dat,
file_name.vhdr, and file_name.vmrk.
}
\description{
Creates an eeg_lst object from BrainVision exported files.The function reads
metadata from the .vhdr BrainVision file, which draws on
the .vmrk and .dat/.eeg files.  All three
files must be in the same directory.
}
\examples{
\dontrun{
# load a single subject
s1 <- read_vhdr("./faces.vhdr", .recording = "1")

# load multiple subjects, extracting subject IDs from file names
faces_list <- lapply(list.files("./", "vhdr"), read_vhdr)
faces <- bind(faces_list)
}

}
\seealso{
Other reading and writing functions:
\code{\link[=read_edf]{read_edf()}},
\code{\link[=read_fif]{read_fif()}},
\code{\link[=read_ft]{read_ft()}},
\code{\link[=read_set]{read_set()}}
}
\concept{reading and writing functions}
