% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{rpink}
\alias{rpink}
\title{Generates pink noise}
\usage{
rpink(n, scale = 1)
}
\arguments{
\item{n}{number of observations.}

\item{scale}{scale parameter.}
}
\value{
a vector
}
\description{
Adapted from Matlab code of Smith (2011)
}
\details{
Reference:
Smith, J.O. Spectral Audio Signal Processing,
\url{http://ccrma.stanford.edu/~jos/sasp/}, online book,
2011 edition.
}
