% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_trans.R
\name{segment}
\alias{segment}
\title{Segments an eegble.}
\usage{
segment(x, ..., lim = c(-1, 1))
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{...}{Description of the event.}
}
\value{
An eegbl.
}
\description{
Segments an eegble.
}
