% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_helpers.R
\name{sig_fft}
\alias{sig_fft}
\title{Compute the one-dimensional discrete Fourier Transform.}
\usage{
sig_fft(x, n = NULL)
}
\arguments{
\item{x}{A vector}

\item{n}{Length of the transformed axis of the output. If \code{n} is smaller
than the length of the input, the input is cropped. If it is larger, the input
is padded with zeros. If n is not given, the length of the input along
the axis specified by axis is used.}
}
\description{
Computes the Discrete Fourier Transform (DFT) of an array
with a fast algorithm, the “Fast Fourier Transform” (FFT). Wrapper
of \link[stats:fft]{stats::fft} with an extra argument similar to python's \href{https://docs.scipy.org/doc/numpy/reference/generated/numpy.fft.fft.html#numpy.fft.fft}{numpy.fft.fft}.
}
\examples{
a <- c(0, 4, 0, 0)
sig_fft(a)
}
