% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_ext.R
\name{slice_signal}
\alias{slice_signal}
\title{Choose samples by the position}
\usage{
slice_signal(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{An eeg_lst object.}

\item{...}{For \code{slice()}: <\code{\link[dplyr:dplyr_data_masking]{data-masking}}> Integer row
values.

Provide either positive values to keep, or negative values to drop.
The values provided must be either all positive or all negative.
Indices beyond the number of rows in the input are silently ignored.

For \code{slice_helpers()}, these arguments are passed on to methods.}

\item{.preserve}{Not in use.}
}
\description{
Choose samples by their ordinal position in the signal table. Grouped eeg_lst object use the ordinal position in the signal table within the group.
}
\seealso{
Other tidyverse-like functions: 
\code{\link{bind}()},
\code{\link{dplyr_verbs}},
\code{\link{drop_incomplete_segments}()}
}
\concept{tidyverse-like functions}
