% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{summaries}
\alias{summaries}
\alias{channel_names}
\alias{channel_ica_names}
\alias{nchannels}
\alias{component_names}
\alias{ncomponents}
\alias{sampling_rate}
\alias{nsamples}
\title{Display information of the eeg_lst object.}
\usage{
channel_names(x, ...)

channel_ica_names(x, ...)

nchannels(x, ...)

component_names(x, ...)

ncomponents(x, ...)

sampling_rate(x, ...)

nsamples(x, ...)
}
\arguments{
\item{x}{An eeg_lst object.}

\item{...}{Not in use.}
}
\description{
\itemize{
\item \verb{*_names()} functions return a vector of names,
\item \verb{n*()} return the number of elements channels or components.
}
}
\details{
Components are only available after running \code{\link[=eeg_ica]{eeg_ica()}}. \code{channel_ica_names()} refers to channels used in the ICA.
}
\seealso{
Other summary functions:
\code{\link[=count_complete_cases_tbl]{count_complete_cases_tbl()}},
\code{\link[=eeg_ica_cor_tbl]{eeg_ica_cor_tbl()}},
\code{\link[=eeg_ica_summary_tbl]{eeg_ica_summary_tbl()}},
\code{\link[=eeg_ica_var_tbl]{eeg_ica_var_tbl()}},
\code{\link[=summary.eeg_lst]{summary.eeg_lst()}}
}
\concept{summary functions}
