% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{summarize_by_id_tbl}
\alias{summarize_by_id_tbl}
\title{Convert an eeg_lst into a summary long-data frame based on a statistics.}
\usage{
summarize_by_id_tbl(x, ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{...}{Other arguments passed on to \code{.funs}. See \link{dplyr-package} help.}
}
\value{
A long tibble.
}
\description{
Convert an eeg_lst into a summary long-data frame based on a statistics.
}
\seealso{
Other summarize: \code{\link{count_complete_cases_tbl}},
  \code{\link{summary}}
}
\concept{summarize}
