% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{summary.eeg_lst}
\alias{summary.eeg_lst}
\title{Summary of eeg_lst information.}
\usage{
\method{summary}{eeg_lst}(object, ...)
}
\arguments{
\item{object}{An eeg_lst object.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
It's also possible to extract the elements by accessing it as a list.
}
\examples{

summary(data_faces_ERPs)
}
\seealso{
Other summary functions:
\code{\link[=count_complete_cases_tbl]{count_complete_cases_tbl()}},
\code{\link[=eeg_ica_cor_tbl]{eeg_ica_cor_tbl()}},
\code{\link[=eeg_ica_summary_tbl]{eeg_ica_summary_tbl()}},
\code{\link[=eeg_ica_var_tbl]{eeg_ica_var_tbl()}},
\code{\link{summaries}}
}
\concept{summary functions}
