% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{summary.eegbl}
\alias{summary.eegbl}
\title{Summary of eegble information.}
\usage{
\method{summary}{eegbl}(object, ...)
}
\arguments{
\item{object}{An eegble object.}

\item{...}{Other options passed to print.tbl for the display of summaries.}
}
\description{
Summary of eegble information.
}
