% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{summary}
\alias{summary}
\alias{channel_names}
\alias{channel_names.signal_tbl}
\alias{channel_names.default}
\alias{channel_names.eeg_lst}
\alias{channel_names.mixing_tbl}
\alias{channel_names.ica_lst}
\alias{nchannels}
\alias{nchannels.default}
\alias{component_names}
\alias{component_names.signal_tbl}
\alias{component_names.eeg_lst}
\alias{ncomponents}
\alias{ncomponents.eeg_lst}
\alias{nsamples}
\alias{nsamples.eeg_lst}
\alias{summary.eeg_lst}
\alias{print.eeg_summary}
\title{Display information of the eeg_lst object.}
\usage{
channel_names(x, ...)

\method{channel_names}{signal_tbl}(x, ...)

\method{channel_names}{default}(x, ...)

\method{channel_names}{eeg_lst}(x, ...)

\method{channel_names}{mixing_tbl}(x, ...)

\method{channel_names}{ica_lst}(x, ...)

nchannels(x, ...)

\method{nchannels}{default}(x, ...)

component_names(x, ...)

\method{component_names}{signal_tbl}(x, ...)

\method{component_names}{eeg_lst}(x, ...)

ncomponents(x, ...)

\method{ncomponents}{eeg_lst}(x, ...)

nsamples(x, ...)

\method{nsamples}{eeg_lst}(x, ...)

\method{summary}{eeg_lst}(object, ...)

\method{print}{eeg_summary}(x, ...)
}
\arguments{
\item{x}{An eeg_lst object.}

\item{...}{additional arguments affecting the summary produced.}

\item{object}{An eeg_lst object.}
}
\description{
\itemize{
\item \code{nchannels()}: Returns the number of channels.
\item \code{channel_names()}: Returns a vector with the name of the channels.
\item \code{nchannels()}: Returns the number of channels.
\item \code{nsamples()}: Returns the number of samples of the recording (or segments).
\item \code{summary()}: Prints a summary of the eeg_lst object. (It also returns a list.)
}
}
\seealso{
Other summarize: \code{\link{count_complete_cases_tbl}}
}
\concept{summarize}
