% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_vhdr}
\alias{write_vhdr}
\title{Write an eeg_lst object to BrainVision file(s) (experimental)}
\usage{
write_vhdr(x, file, overwrite = FALSE)
}
\arguments{
\item{x}{\code{eeg_lst} object.}

\item{file}{Either a vector of file names (with the same length as the number of recordings), one file name (which will be appended the recording name if needed), or a path which will use the recording names as file names. By default, it will use the recording name in the current path as file names.}

\item{overwrite}{Stop writing if the file exists.}
}
\description{
The function will write each recording in a different file. The function is highly experimental, if a file fails to open in BrainVision, please report the bug in (https://github.com/bnicenboim/eeguana/issues).
}
