% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotate_electrodes}
\alias{annotate_electrodes}
\title{Adds the electrode labels to a head shape}
\usage{
annotate_electrodes(.label = .key, ...)
}
\arguments{
\item{.label}{Label name for the electrodes}

\item{...}{other arguments to control the text}
}
\value{
A layer for a ggplot
}
\description{
Adds the electrode labels to a head shape
}
\examples{
library(ggplot2)
data_faces_ERPs \%>\%
  eeg_filter(between(as_time(.sample, .unit = "milliseconds"), 100, 200)) \%>\%
  eeg_group_by(condition) \%>\%
  eeg_summarize(across_ch(mean, na.rm = TRUE)) \%>\%
  plot_topo() +
  annotate_head(size = .9, color = "black", stroke = 1) +
  annotate_electrodes(color = "gray")
}
\seealso{
Other plotting functions:
\code{\link[=annotate_events]{annotate_events()}},
\code{\link[=annotate_head]{annotate_head()}},
\code{\link[=eeg_downsample]{eeg_downsample()}},
\code{\link[=ggplot.eeg_lst]{ggplot.eeg_lst()}},
\code{\link[=plot.eeg_lst]{plot.eeg_lst()}},
\code{\link[=plot_components]{plot_components()}},
\code{\link[=plot_in_layout]{plot_in_layout()}},
\code{\link[=plot_topo]{plot_topo()}},
\code{\link[=theme_eeguana]{theme_eeguana()}}
}
\concept{plotting functions}
