% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotate_events}
\alias{annotate_events}
\title{Adds a layer with the events on top of a plot of an eeg_lst.}
\usage{
annotate_events(data = NULL, alpha = 0.2)
}
\arguments{
\item{data}{The data to be displayed in this layer. There are three options:
* If NULL, the default, the events table is inherited from the plot data as specified
in the call to ggplot().
* An events table will override the plot events table data.}

\item{alpha}{new alpha level in 0,1.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot}} layer.
}
\description{
Adds a layer with the events on top of a plot of an eeg_lst.
}
\seealso{
Other plotting functions: \code{\link{annotate_head}},
  \code{\link{eeg_downsample}},
  \code{\link{ggplot.eeg_lst}}, \code{\link{plot.eeg_lst}},
  \code{\link{plot_components}},
  \code{\link{plot_in_layout}}, \code{\link{plot_topo}},
  \code{\link{theme_eeguana}}
}
\concept{plotting functions}
