% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotate_head}
\alias{annotate_head}
\title{Annotates a head in a ggplot}
\usage{
annotate_head(size = 1.1, color = "black", stroke = 1)
}
\arguments{
\item{size}{Size of the head}

\item{color}{Color of the head}

\item{stroke}{Line thickness}
}
\value{
A layer for a ggplot
}
\description{
Annotates a head in a ggplot
}
