% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{annotate_head}
\alias{annotate_head}
\title{Add a head shape to a ggplot}
\usage{
annotate_head(size = 1.1, color = "black", stroke = 1)
}
\arguments{
\item{size}{Size of the head}

\item{color}{Color of the head}

\item{stroke}{Line thickness}
}
\value{
A layer for a ggplot
}
\description{
Adds the outline of a head and nose to a ggplot.
}
\examples{

data_faces_ERPs \%>\% 
    filter(between(as_time(.sample_id, unit = "milliseconds"),100,200)) \%>\% 
    group_by(condition) \%>\%
    summarize_all_ch(mean, na.rm = TRUE) \%>\%
    plot_topo() +
    annotate_head(size = 1, color = "black", stroke = 2)


}
