% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as.data.frame.eeg_lst}
\alias{as.data.frame.eeg_lst}
\title{Convert an eeg_lst to a (base) data frame.}
\usage{
\method{as.data.frame}{eeg_lst}(x, .unit = "s")
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{.unit}{Unit for the \code{.time} column of the transformed object: "s" (default), "ms", "samples".}
}
\value{
A data frame.
}
\description{
Convert an eeg_lst to a (base) data frame.
}
