% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as.data.frame.eeg_lst}
\alias{as.data.frame.eeg_lst}
\title{Convert an eeg_lst to a (base) data frame.}
\usage{
\method{as.data.frame}{eeg_lst}(...)
}
\arguments{
\item{...}{Other arguments passed on to individual methods.}
}
\value{
A tibble.
}
\description{
Convert an eeg_lst to a (base) data frame.
}
\seealso{
Other tibble: 
\code{\link{as_tibble.eeg_lst}()},
\code{\link{as_tibble.psd_lst}()}
}
\concept{tibble}
