% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as.data.frame.psd_lst}
\alias{as.data.frame.psd_lst}
\title{Convert a psd_lst to a (base) data frame.}
\usage{
\method{as.data.frame}{psd_lst}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A \code{psd_lst} object.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{...}{Additional arguments to be passed to or from other methods.}
}
\value{
A data.frame.
}
\description{
Convert a psd_lst to a (base) data frame.
}
\seealso{
Other tibble: 
\code{\link{as_tibble.eeg_lst}()},
\code{\link{as_tibble.psd_lst}()}
}
\concept{tibble}
