% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as.data.frame.psd_lst}
\alias{as.data.frame.psd_lst}
\title{Convert a psd_lst to a (base) data frame.}
\usage{
\method{as.data.frame}{psd_lst}(x)
}
\arguments{
\item{x}{A \code{psd_lst} object.}
}
\value{
A tibble.
}
\description{
Convert a psd_lst to a (base) data frame.
}
\seealso{
Other tibble: 
\code{\link{as_tibble.eeg_lst}()},
\code{\link{as_tibble.psd_lst}()}
}
\concept{tibble}
