% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{as_channel_dbl}
\alias{as_channel_dbl}
\title{Coerce a vector of real (double) numbers into a channel object}
\usage{
as_channel_dbl(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
A channel_dbl.
}
\description{
Coerce a vector of real (double) numbers into a channel object
}
\seealso{
Other channel: 
\code{\link{channel_dbl}()},
\code{\link{is_channel_dbl}()}
}
\concept{channel}
