% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_tbl.R
\name{as_tibble.eeg_lst}
\alias{as_tibble.eeg_lst}
\alias{as_data_frame.eeg_lst}
\alias{as.tibble.eeg_lst}
\title{Convert an eeg_lst to a wide table.}
\usage{
\method{as_tibble}{eeg_lst}(x, add_segments = TRUE,
  add_channels_info = TRUE)

\method{as_data_frame}{eeg_lst}(x, add_segments = TRUE,
  add_channels_info = TRUE)

as.tibble.eeg_lst(x, add_segments = TRUE, add_channels_info = TRUE)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{add_segments}{Whether the segments table

\code{as_data_frame} and \code{as.tibble} are aliases.}
}
\value{
A tibble.
}
\description{
Convert the signal_tbl table from wide to long format, and optionally \code{left_join}s the segment table
}
\seealso{
Other tibble: \code{\link{as.data.frame.eeg_lst}}
}
\concept{tibble}
