% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_trans.R
\name{as_tibble.eegbl}
\alias{as_tibble.eegbl}
\title{Convert an eegble to a tibble.}
\usage{
as_tibble.eegbl(x, ..., chans, .id = "id")
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{...}{Other arguments passed on to individual methods.}

\item{chans}{Channell names to be used.}

\item{.id}{Name of the column that identifies the data.}
}
\value{
A tibble.
}
\description{
Convert an eegble to a tibble.
}
