% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_trans.R
\name{baseline}
\alias{baseline}
\title{Baseline an eegble.}
\usage{
baseline(x, ...)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{t}{A negative number indicating from when to baseline
(TO COMPLETE). The default is to use all the negative times.}
}
\value{
An eegbl.
}
\description{
Baseline an eegble.
}
