% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exported_utils.R
\name{between}
\alias{between}
\title{Convenience function for range subsets}
\description{
\code{between} is a thin wrapper for the between function of \link[data.table:between]{data.table::between}.
It is equivalent to x >= lower & x <= upper when incbounds=TRUE, or x > lower & y < upper when FALSE. In comparison
with \link[dplyr:between]{dplyr::between}, it doesn't loose the class of its argument, and it's more appropriate for manipulating
the column \code{.sample}. For more information and the description of the arguments, see \link[data.table:between]{data.table::between}.
}
\examples{

data_faces_ERPs \%>\%
  eeg_filter(.sample \%>\% between(10, 100))

}
