% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_transformation.R
\name{bind}
\alias{bind}
\title{Bind eegble objects.}
\usage{
bind(...)
}
\arguments{
\item{...}{Eegble objects to combine.}
}
\value{
An \code{eegble} object.
}
\description{
Binds eegble and throws a warning if there is a mismatch in the channel information.
}
