% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_transformation.R
\name{ch_baseline}
\alias{ch_baseline}
\title{Baseline an eeg_lst}
\usage{
ch_baseline(x, ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{...}{Not in use.}

\item{time}{A negative number indicating from when to baseline; the interval is defined as \link{time,0}. The default is to use all the negative times.}

\item{sample_id}{A negative number indicating from when to baseline. The default is to use all the negative times. (\code{time} is ignored if  sample_id is used).}
}
\value{
An eeg_lst.
}
\description{
Subtract the average or baseline of the points in a defined interval from all points in the segment.
}
\seealso{
Other channel: \code{\link{ch_rereference}},
  \code{\link{channel_dbl}}, \code{\link{chs_mean}},
  \code{\link{is_channel_dbl}}
}
\concept{channel}
