% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{ch_rereference}
\alias{ch_rereference}
\alias{ch_rereference.channel_dbl}
\alias{ch_rereference.eeg_lst}
\title{Rereference a channel or group of channels.}
\usage{
ch_rereference(x, ..., na.rm = FALSE)

\method{ch_rereference}{channel_dbl}(x, ..., na.rm = FALSE)

\method{ch_rereference}{eeg_lst}(x, ..., na.rm = FALSE, exclude = NULL)
}
\arguments{
\item{x}{A channel to be referenced an eeg_lst where all the channels will be re-referenced (except for the ones in exclude).}

\item{...}{Channels that will be averaged as the reference.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}

\item{exclude}{A character vector of channels to exclude from referencing.}
}
\value{
A rereferenced channel or an eeg_lst with all channels re-referenced.
}
\description{
Rereference a channel or group of channels.
}
\details{
Notice that this function will update the channels one by one when used inside a mutate and not all at the same time.
}
\examples{
\dontrun{
# Rereference all channels used the linked mastoids (average of the two mastoids)

faces_segs \%>\% ch_rereference(M1, M2)
}
}
\seealso{
Other channel: \code{\link{ch_baseline}},
  \code{\link{channel_dbl}}, \code{\link{chs_mean}},
  \code{\link{is_channel_dbl}}
}
\concept{channel}
