% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{channel_dbl}
\alias{channel_dbl}
\title{Builds a channel.}
\usage{
channel_dbl(values, x = NA_real_, y = NA_real_, z = NA_real_,
  reference = NA, ...)
}
\arguments{
\item{values}{Vector of doubles indicating amplitudes.}

\item{x}{Position in the scalp.}

\item{y}{Position in the scalp.}

\item{z}{Position in the scalp.}

\item{reference}{Reference electrode.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Builds a channel.
}
\examples{

Cz <- channel_dbl(runif(100,-5,5))
}
\seealso{
Other channel: \code{\link{ch_baseline}},
  \code{\link{ch_rereference}}, \code{\link{chs_mean}},
  \code{\link{is_channel_dbl}}
}
\concept{channel}
