% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{channels_tbl}
\alias{channels_tbl}
\alias{channels_tbl<-}
\title{Functions to get or set the channel information of an eeg_lst object.}
\usage{
channels_tbl(.data, ...)

channels_tbl(.data) <- value
}
\arguments{
\item{.data}{An eeg_lst object.}

\item{...}{Not in use.}

\item{value}{A channel table.}
}
\value{
A table.
}
\description{
Functions to get or set the channel information of an eeg_lst object.
}
\examples{
\dontrun{
library(dplyr)
# Get channel information:
channels_tbl(faces_seg)
# Set channel information using dplyr's mutate and replace
channels_tbl(faces_seg) <- mutate(channels_tbl(faces_seg),
              .channel = replace(.channel, .channel=="HEOG", "EOGH"),
              .channel = replace(.channel, .channel=="VEOG", "EOGV"))
}


}
\seealso{
Other functions to expose internal parts of eeg_(ica_)_lst: 
\code{\link{events_tbl}()},
\code{\link{ica_matrix_lst}()},
\code{\link{segments_tbl}()},
\code{\link{signal_tbl}()}
}
\concept{functions to expose internal parts of eeg_(ica_)_lst}
