% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/editing.R
\name{channels_tbl}
\alias{channels_tbl}
\alias{channels_tbl.eeg_lst}
\alias{channels_tbl<-}
\alias{channels_tbl<-.eeg_lst}
\title{Functions to get or set the channel information of an eeg_lst object.}
\usage{
channels_tbl(x, ...)

\method{channels_tbl}{eeg_lst}(x, ...)

channels_tbl(x) <- value

\method{channels_tbl}{eeg_lst}(x) <- value
}
\arguments{
\item{x}{An eeg_lst object.}

\item{...}{Not in use.}

\item{value}{A channel table.}
}
\value{
A table.
}
\description{
Functions to get or set the channel information of an eeg_lst object.
}
