% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{channels_tbl}
\alias{channels_tbl}
\alias{channels_tbl<-}
\title{Functions to get or set the channel information of an eeg_lst object.}
\usage{
channels_tbl(.data, ...)

channels_tbl(.data) <- value
}
\arguments{
\item{.data}{An eeg_lst object.}

\item{...}{Not in use.}

\item{value}{A channel table.}
}
\value{
A table.
}
\description{
Functions to get or set the channel information of an eeg_lst object.
}
