% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{chs_fun}
\alias{chs_fun}
\title{Get the by-sample (or by-row) function of the specified channels.}
\usage{
chs_fun(x, .funs, ..., pars = list())
}
\arguments{
\item{x}{An eeg_lst.}

\item{.funs}{A function \code{fun}, a quosure style lambda \code{~ fun(.)} or a list of either form.}

\item{...}{A channel or a group of unquoted or quoted channels (if an \code{eeg_lst} is not specified).}

\item{pars}{List that contains the additional arguments for the function calls in .funs.}
}
\value{
A new channel or an \code{eeg_lst} object with a channel where a function was instead of the previous channels.
}
\description{
Get the by-sample (or by-row) function of the specified channels.
}
\seealso{
Other channel functions: 
\code{\link{chs_mean}()}
}
\concept{channel functions}
