% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{chs_fun}
\alias{chs_fun}
\alias{chs_fun.channel_dbl}
\alias{chs_fun.character}
\alias{chs_fun.eeg_lst}
\title{Get the by-sample (or by-row) function of the specified channels.}
\usage{
chs_fun(x, .funs, ...)

\method{chs_fun}{channel_dbl}(..., .funs, pars = list())

\method{chs_fun}{character}(..., .funs, pars = list())

\method{chs_fun}{eeg_lst}(x, .funs, pars = list(), ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{.funs}{List of function calls generated by \code{\link[=funs]{funs()}}, or a
character vector of function names, or simply a function.

Bare formulas are passed to \code{\link[rlang:as_function]{rlang::as_function()}} to create
purrr-style lambda functions. Note that these lambda prevent
hybrid evaluation from happening and it is thus more efficient to
supply functions like \code{mean()} directly rather than in a
lambda-formula.}

\item{...}{A channel or a group of unquoted or quoted channels (if an \code{eeg_lst} is not specified).}

\item{pars}{list that contains the additional arguments for the function calls in .funs.}

\item{pars}{list that contains the additional arguments for the function calls in .funs.}
}
\value{
A new channel or an \code{eeg_lst} object with a channel where a function was instead of the previous channels.
}
\description{
Get the by-sample (or by-row) function of the specified channels.
}
