% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{chs_mean}
\alias{chs_mean}
\title{Get the by-sample (or by-row) mean of the specified channels.}
\usage{
chs_mean(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An \R object.  Currently there are methods for
    numeric/logical vectors and \link[=Dates]{date},
    \link{date-time} and \link{time interval} objects.  Complex vectors
    are allowed for \code{trim = 0}, only.}

\item{...}{A channel or a group of channels, or an \code{eeg_lst} object.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
A new channel or an \code{eeg_lst} object with a \code{mean} channel instead of the previous channels.
}
\description{
Wrapper of \code{rowMeans} that performs a by-sample mean of the specified channels.
}
\examples{
\dontrun{

faces_segs_some \%>\%
               transmute(Occipital = chs_mean(O1, O2, Oz, na.rm = TRUE),
                         Parietal = chs_mean(P3, P4, P7,  P8, Pz, na.rm = TRUE))

faces_segs_some \%>\%
               chs_mean(na.rm = TRUE)
}
}
\seealso{
Other channel: \code{\link{ch_baseline}},
  \code{\link{ch_rereference}}, \code{\link{channel_dbl}},
  \code{\link{is_channel_dbl}}
}
\concept{channel}
