% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{chs_mean}
\alias{chs_mean}
\title{Get the by-sample (or by-row) mean of the specified channels.}
\usage{
chs_mean(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{...}{A channel or a group of unquoted or quoted channels (if an \code{eeg_lst} is not specified).}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
A new channel or an \code{eeg_lst} object with a \code{mean} channel instead of the previous channels.
}
\description{
Wrapper of \code{rowMeans} that performs a by-sample mean of the specified channels.
}
\examples{
\dontrun{

faces_segs_some \%>\%
  transmute(
    Occipital = chs_mean(O1, O2, Oz, na.rm = TRUE),
    Parietal = chs_mean(P3, P4, P7, P8, Pz, na.rm = TRUE)
  )

faces_segs_some \%>\%
  chs_mean(na.rm = TRUE)
}
}
\seealso{
Other channel functions: \code{\link{chs_fun}}
}
\concept{channel functions}
