% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{chs_mean}
\alias{chs_mean}
\title{Get the by-sample (or by-row) mean of the specified channels.}
\usage{
chs_mean(x, ..., na.rm = FALSE)
}
\arguments{
\item{...}{A channel or a group of channels, or an \code{eeg_lst} object.}

\item{na.rm}{}
}
\value{
A new channel or an \code{eeg_lst} object with a \code{mean} channel instead of the previous channels.
}
\description{
Wrapper of \code{rowMeans} that performs a by-sample mean of the specified channels.
}
\seealso{
Other channel: \code{\link{ch_baseline}},
  \code{\link{ch_rereference}}, \code{\link{channel_dbl}},
  \code{\link{is_channel_dbl}}
}
\concept{channel}
