% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{component_dbl}
\alias{component_dbl}
\title{Builds a component.}
\usage{
component_dbl(values)
}
\arguments{
\item{values}{Vector of doubles indicating amplitudes.}
}
\description{
Builds a component.
}
\examples{

Cz <- component_dbl(runif(100,-5))
}
\seealso{
Other component: \code{\link{is_component_dbl}}
}
\concept{component}
