% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{count_complete_cases_tbl}
\alias{count_complete_cases_tbl}
\title{Count number of complete segments of an eeg_lst object.}
\usage{
count_complete_cases_tbl(x, ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{...}{Variables from the segment table to group by.}
}
\value{
A tbl.
}
\description{
Count number of complete segments of an eeg_lst object.
}
\examples{
\dontrun{
faces_segs_some \%>\% count_complete_cases(.recording, .description)
}
}
\seealso{
Other summary functions: 
\code{\link{eeg_ica_cor_tbl}()},
\code{\link{eeg_ica_summary_tbl}()},
\code{\link{eeg_ica_var_tbl}()},
\code{\link{summaries}},
\code{\link{summary.eeg_lst}()}
}
\concept{summarize}
\concept{summary functions}
