% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{count_complete_cases_tbl}
\alias{count_complete_cases_tbl}
\title{Count number of complete segments of an eeg_lst object.}
\usage{
count_complete_cases_tbl(x, ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{...}{Variables to group by.}
}
\value{
A tbl.
}
\description{
Count number of complete segments of an eeg_lst object.
}
\examples{
\dontrun{

faces_segs_some \%>\% count_complete_cases(recording, description)
}
}
\seealso{
Other summarize: \code{\link{summary}}

Other summarize: \code{\link{summary}}
}
\concept{summarize}
