% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{drop_incomplete_segments}
\alias{drop_incomplete_segments}
\title{Drop segments with NAs from the eeg_lst}
\usage{
drop_incomplete_segments(x)
}
\arguments{
\item{x}{eeg_lst}
}
\value{
An eeg_lst object
}
\description{
Drop segments with NAs from the eeg_lst.
}
\seealso{
Other tidyverse-like functions: 
\code{\link{dplyr_verbs}},
\code{\link{eeg_bind}()},
\code{\link{eeg_slice_signal}()}
}
\concept{tidyverse-like functions}
