% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_edition.R
\name{eeg_artif}
\alias{eeg_artif}
\alias{eeg_artif_minmax}
\alias{eeg_artif_minmax.eeg_lst}
\alias{eeg_artif_step}
\alias{eeg_artif_step.eeg_lst}
\alias{eeg_artif_amplitude}
\alias{eeg_artif_amplitude.eeg_lst}
\title{Detect artifacts and add them in the events table of an eeg_lst.}
\usage{
eeg_artif_minmax(.data, ..., threshold = 100, window = 0.2,
  lim = c(-window, window), unit = "s")

\method{eeg_artif_minmax}{eeg_lst}(.data, ..., threshold = 100,
  window = 0.2, lim = c(-window, window), unit = "s")

eeg_artif_step(.data, ..., threshold = 50, window = 0.2,
  lim = c(-window, window), unit = "s")

\method{eeg_artif_step}{eeg_lst}(.data, ..., threshold = 50,
  window = 0.2, lim = c(-window, window), unit = "s")

eeg_artif_amplitude(.data, ..., threshold = c(-200, 200), lim = c(-0.2,
  0.2), unit = "s")

\method{eeg_artif_amplitude}{eeg_lst}(.data, ..., threshold = c(-200,
  200), lim = c(-0.2, 0.2), unit = "s")
}
\arguments{
\item{.data}{An \code{eeg_lst} object.}

\item{...}{Channels to include. All the channels by default, but eye channels should be removed.}

\item{threshold}{Voltage threshold that indicates an artifact. This is between two consecutive data points for \code{eeg_artif_step}, and in a \code{window} for \code{eeg_artif_minmax}.}

\item{window}{Sliding window length for the artifact detection (same unit as \code{lim}).}

\item{lim}{Vector indicating the time before and after the artifact that will be included in events_tbl (by default the size the window before and afterwards).}

\item{unit}{"seconds" (or "s"), "milliseconds" (or "ms")}
}
\value{
An \code{eeg_lst}.
}
\description{
\itemize{
\item \code{eeg_artif_step()}
\itemize{
\item \code{eeg_artif_minmax()}
}
}
}
\details{
\code{eeg_artif_minmax()}and \code{eeg_artif_step()} can be used to detect blinks and horizontal eye movements
in the electro-oculographic (V/HEOG) channels or large voltage jumps in other channels.
For the EOG channels, a relatively low threshold (e.g., 30 µV) is recommended. For non EOG channels, a relatively high threshold (e.g., 100 µV) would be
more appropriate.
}
