% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{eeg_baseline}
\alias{eeg_baseline}
\title{Baseline an eeg_lst}
\usage{
eeg_baseline(x, ..., .lim = -Inf, .unit = "s")
}
\arguments{
\item{x}{An \code{eeg_lst} object or a channel.}

\item{...}{Channels to include. All channels by default.}

\item{.lim}{A negative number indicating from when to baseline; the interval is defined as [lim,0]. The default is to use all the negative times.}

\item{.unit}{"seconds" (or "s"), "milliseconds" (or "ms")}
}
\value{
An eeg_lst.
}
\description{
Subtract the average or baseline of the points in a defined interval from all points in the segment.
}
\seealso{
Other preprocessing functions:
\code{\link[=eeg_downsample]{eeg_downsample()}},
\code{\link[=eeg_ica]{eeg_ica()}},
\code{\link[=eeg_ica_keep]{eeg_ica_keep()}},
\code{\link[=eeg_rereference]{eeg_rereference()}},
\code{\link[=eeg_segment]{eeg_segment()}},
\code{\link{filt}}
}
\concept{preprocessing functions}
