% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_ext.R
\name{eeg_bind}
\alias{eeg_bind}
\alias{bind}
\title{Bind eeg_lst objects.}
\usage{
eeg_bind(...)

bind(...)
}
\arguments{
\item{...}{\code{eeg_lst} objects to combine.}
}
\value{
An \code{eeg_lst} object.
}
\description{
Binds eeg_lst and throws a warning if there is a mismatch in the channel information.
}
\examples{
\dontrun{

# Load multiple subjects using purrr::map, extracting subject IDs from file names.
faces_list <- purrr::map(list.files("./", "vhdr"), ~
read_vhdr(.x))
# Bind all the eeg_lsts into a large one:
faces <- eeg_bind(faces_list)
}

}
\seealso{
Other tidyverse-like functions: 
\code{\link{dplyr_verbs}},
\code{\link{drop_incomplete_segments}()},
\code{\link{eeg_slice_signal}()}
}
\concept{tidyverse-like functions}
