% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_processing.R
\name{eeg_downsample}
\alias{eeg_downsample}
\title{Downsample EEG data}
\usage{
eeg_downsample(
  .data,
  .q = 2,
  .max_sample = NULL,
  .n = if (.ftype == "iir") 8 else 30,
  .ftype = "iir",
  .multiple_times = FALSE,
  ...
)
}
\arguments{
\item{.data}{An eeg_lst object.}

\item{.q}{integer factor(s) to downsample by.}

\item{.max_sample}{Optionally, the (approximated) maximum sample number can be defined here, which is at least half of the total number of samples}

\item{.n}{Order of the filter used prior to the downsampling, specified as a
positive integer. Default: 8 if \code{ftype} equals \code{"iir"}; 30 of
\code{ftype} equals \code{"fir"}.}

\item{.ftype}{filter type; either \code{"fir"}, specifying a FIR filter of
length \code{n} designed with the function \code{\link[gsignal]{fir1}}, or
\code{"iir"} (default), specifying an IIR Chebyshev filter of order 8 using
the function \code{\link[gsignal]{cheby1}}.}

\item{.multiple_times}{Indicates whether to factorize `q`` and apply the downsampling in steps.}

\item{...}{Not in use.}
}
\description{
Downsample a signal_tbl by a factor \code{q}, using an FIR or IIR filter.
This is a wrapper for \code{\link[gsignal:decimate]{decimate}} from the
\code{gsignal} package, see its documentation for details.
}
\details{
A factor q larger than 13 can result in NAs. To avoid this,
the downsampling can be done in steps. For example, instead of setting
\code{q = 20}, it is possible to set \code{q = c(2,10)}.
}
\seealso{
Other preprocessing functions:
\code{\link[=eeg_baseline]{eeg_baseline()}},
\code{\link[=eeg_ica]{eeg_ica()}},
\code{\link[=eeg_ica_keep]{eeg_ica_keep()}},
\code{\link[=eeg_rereference]{eeg_rereference()}},
\code{\link[=eeg_segment]{eeg_segment()}},
\code{\link{filt}}

Other plotting functions:
\code{\link[=annotate_electrodes]{annotate_electrodes()}},
\code{\link[=annotate_events]{annotate_events()}},
\code{\link[=annotate_head]{annotate_head()}},
\code{\link[=browse_ica]{browse_ica()}},
\code{\link[=ggplot.eeg_lst]{ggplot.eeg_lst()}},
\code{\link[=plot.eeg_lst]{plot.eeg_lst()}},
\code{\link[=plot_components]{plot_components()}},
\code{\link[=plot_in_layout]{plot_in_layout()}},
\code{\link[=plot_topo]{plot_topo()}},
\code{\link[=theme_eeguana]{theme_eeguana()}}
}
\concept{plotting functions}
\concept{preprocessing functions}
