% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_processing.R
\name{eeg_downsample}
\alias{eeg_downsample}
\title{Downsample EEG data}
\usage{
eeg_downsample(x, q = 2, max_sample = NULL, ...)
}
\arguments{
\item{x}{An eeg_lst object.}

\item{q}{integer factor(s) to downsample by.}

\item{max_sample}{Optionally, the (approximated) maximum sample number can be defined here, which is at least half of the total number of samples.}

\item{...}{Not in use.}
}
\description{
Downsample a signal_tbl by a factor \code{q}, using an FIR or IIR filter.
This is a wrapper for \code{\link[signal:decimate]{decimate}} from the
\code{\link[signal:signal]{signal}} package, see its documentation for details. Notice that
the code of the signal package might be outdated.
}
\details{
A factor q larger than 13 can result in NAs. To avoid this,
the downsampling can be done in steps. For example, instead of setting
\code{q = 20}, it is possible to set \code{q = c(2,10)}.
}
\concept{eeg}
