% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_edition.R
\name{eeg_events_to_NA}
\alias{eeg_events_to_NA}
\alias{eeg_events_to_NA.eeg_lst}
\title{Remove (transform to NA) problematic events from an eeg_lst.}
\usage{
eeg_events_to_NA(x, ...)

\method{eeg_events_to_NA}{eeg_lst}(x, ..., all_chans = FALSE,
  entire_seg = TRUE, drop_events = TRUE)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{...}{Description of the problematic event.}

\item{all_chans}{If set to \code{TRUE},
it will consider samples from all channels (Default:  \code{all_chans = FALSE}).}

\item{entire_seg}{If set to \code{FALSE}, it will consider only the marked part of the segment,
otherwise it will consider the entire segment (Default: entire_seg = TRUE).}

\item{drop_events}{If set to \code{TRUE} (default), the events that were used for setting signals to NA, will be removed from the events table.}
}
\value{
An eeg_lst.
}
\description{
Remove (transform to NA) problematic events from an eeg_lst.
}
