% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifact_edition.R
\name{eeg_events_to_NA}
\alias{eeg_events_to_NA}
\title{Remove (transform to NA) problematic events from the signal table of an eeg_lst.}
\usage{
eeg_events_to_NA(
  x,
  ...,
  all_chs = FALSE,
  entire_seg = TRUE,
  drop_events = TRUE
)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{...}{Description of the problematic event.}

\item{all_chs}{If set to \code{TRUE},
it will consider samples from all channels (Default:  \code{all_chs = FALSE}).}

\item{entire_seg}{If set to \code{FALSE}, it will consider only the marked part of the segment,
otherwise it will consider the entire segment (Default: entire_seg = TRUE).}

\item{drop_events}{If set to \code{TRUE} (default), the events that were used for setting signals to NA, will be removed from the events table.}
}
\value{
An eeg_lst.
}
\description{
Remove (transform to NA) problematic events from the signal table of an eeg_lst.
}
\seealso{
Other events functions: 
\code{\link{eeg_artif}}
}
\concept{events functions}
