% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{eeg_ica}
\alias{eeg_ica}
\alias{eeg_ica.eeg_lst}
\title{Apply ICA}
\usage{
eeg_ica(.data, ...)

\method{eeg_ica}{eeg_lst}(.data, ..., method = "fastica",
  config = list())
}
\arguments{
\item{.data}{}

\item{...}{}

\item{method}{}

\item{config}{}

\item{.data}{}

\item{...}{Channels for the eeg_ica}
}
\value{
An ica_lst object
}
\description{
Apply ICA
}
