% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{eeg_ica_cor_tbl}
\alias{eeg_ica_cor_tbl}
\title{Show correlations between ICA sources and eye (EOG) electrodes.}
\usage{
eeg_ica_cor_tbl(.data, ...)
}
\arguments{
\item{.data}{An \code{eeg_ica_lst} object}

\item{...}{If left empty, the function will assume that EOG channels include eog/EOG in their names, alternatively, EOG channels can be selected here.}
}
\value{
A table with the correlations between each component and each EOG channel in each recording.
}
\description{
Show correlations between ICA sources and eye (EOG) electrodes.
}
\seealso{
Other ICA functions: 
\code{\link{eeg_ica}()},
\code{\link{eeg_ica_keep}()},
\code{\link{eeg_ica_show}()},
\code{\link{eeg_ica_summary_tbl}()},
\code{\link{eeg_ica_var_tbl}()},
\code{\link{plot_components}()}

Other summary functions: 
\code{\link{count_complete_cases_tbl}()},
\code{\link{eeg_ica_summary_tbl}()},
\code{\link{eeg_ica_var_tbl}()},
\code{\link{summaries}},
\code{\link{summary.eeg_lst}()}
}
\concept{ICA functions}
\concept{summary functions}
