% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{eeg_ica_keep}
\alias{eeg_ica_keep}
\title{Select independent components (or sources) to keep.}
\usage{
eeg_ica_keep(.data, ...)
}
\arguments{
\item{.data}{An eeg_ica_lst object}

\item{...}{Components to keep from the mixing matrix of the ICA transformation. See \link[dplyr:select]{dplyr::select} and \link[tidyselect:select_helpers]{tidyselect::select_helpers} for details.}
}
\description{
This function will transform the channels according to the indepent components that are kept or removed.
}
\seealso{
Other preprocessing functions: 
\code{\link{eeg_baseline}()},
\code{\link{eeg_downsample}()},
\code{\link{eeg_ica}()},
\code{\link{eeg_rereference}()},
\code{\link{eeg_segment}()},
\code{\link{filt}}

Other ICA functions: 
\code{\link{eeg_ica_cor_tbl}()},
\code{\link{eeg_ica_show}()},
\code{\link{eeg_ica_summary_tbl}()},
\code{\link{eeg_ica_var_tbl}()},
\code{\link{eeg_ica}()},
\code{\link{plot_components}()}
}
\concept{ICA functions}
\concept{preprocessing functions}
