% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ica.R
\name{eeg_ica_show}
\alias{eeg_ica_show}
\title{Add independent components (or sources) to the signal table for visualization.}
\usage{
eeg_ica_show(.data, ...)
}
\arguments{
\item{.data}{An eeg_ica_lst object}

\item{...}{Components to extract from the mixing matrix of the ICA transformation.}
}
\description{
Add independent components (or sources) to the signal table for visualization.
}
\seealso{
Other ICA functions: \code{\link{eeg_ica_cor_tbl}},
  \code{\link{eeg_ica_keep}},
  \code{\link{eeg_ica_summary_tbl}},
  \code{\link{eeg_ica_var_tbl}}, \code{\link{eeg_ica}},
  \code{\link{plot_components}}
}
\concept{ICA functions}
