% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{eeg_ica_summary_tbl}
\alias{eeg_ica_summary_tbl}
\title{Show a table with a summary of the results of the ICA.}
\usage{
eeg_ica_summary_tbl(.data, ...)
}
\arguments{
\item{.data}{An \code{eeg_ica_lst} object}

\item{...}{If left empty, the function will assume that EOG channels include eog/EOG in their names, alternatively, EOG channels can be selected here.}
}
\value{
A table with the variance explained by each component, and the correlation between EOG channels and each components in each recording.
}
\description{
This function generates a table with the variance explained by each ICA component, and the correlations
between ICA components and EOG channels. See more details in \link{eeg_ica_cor_tbl} and \link{eeg_ica_var_tbl}.
}
\seealso{
Other ICA functions: 
\code{\link{eeg_ica}()},
\code{\link{eeg_ica_cor_tbl}()},
\code{\link{eeg_ica_keep}()},
\code{\link{eeg_ica_show}()},
\code{\link{eeg_ica_var_tbl}()},
\code{\link{plot_components}()}

Other summary functions: 
\code{\link{count_complete_cases_tbl}()},
\code{\link{eeg_ica_cor_tbl}()},
\code{\link{eeg_ica_var_tbl}()},
\code{\link{summaries}},
\code{\link{summary.eeg_lst}()}
}
\concept{ICA functions}
\concept{summary functions}
