% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{eeg_ica_var_tbl}
\alias{eeg_ica_var_tbl}
\title{Show the variance explained for each ICA sources.}
\usage{
eeg_ica_var_tbl(.data, ..., .max_sample = 1e+05)
}
\arguments{
\item{.data}{An \code{eeg_ica_lst} object}

\item{...}{Not in use.}

\item{.max_sample}{The maximum number of samples to use for calculating the variance explained.}
}
\value{
A table with the variance explained by each component in each recording.
}
\description{
This function shows the variance explained by each ICA component following the approach of the matlab function ....
}
\details{
If the dataset is large, this function can take very long to run. Setting a maximum number of samples (\code{max_sample}) will speed up the calculations by downsampling the data.
}
\seealso{
Other ICA functions:
\code{\link[=eeg_ica]{eeg_ica()}},
\code{\link[=eeg_ica_cor_tbl]{eeg_ica_cor_tbl()}},
\code{\link[=eeg_ica_keep]{eeg_ica_keep()}},
\code{\link[=eeg_ica_show]{eeg_ica_show()}},
\code{\link[=eeg_ica_summary_tbl]{eeg_ica_summary_tbl()}},
\code{\link[=plot_components]{plot_components()}}

Other summary functions:
\code{\link[=count_complete_cases_tbl]{count_complete_cases_tbl()}},
\code{\link[=eeg_ica_cor_tbl]{eeg_ica_cor_tbl()}},
\code{\link[=eeg_ica_summary_tbl]{eeg_ica_summary_tbl()}},
\code{\link{summaries}},
\code{\link[=summary.eeg_lst]{summary.eeg_lst()}}
}
\concept{ICA functions}
\concept{summary functions}
