% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{eeg_lst}
\alias{eeg_lst}
\title{Builds an eeg_lst.}
\usage{
eeg_lst(signal = NULL, events = NULL, segments = NULL)
}
\arguments{
\item{events}{}

\item{segments}{}

\item{signal_tbl}{}
}
\value{
A valid eggble.
}
\description{
Builds an eeg_lst.
}
\seealso{
Other eeg_lst: \code{\link{is_eeg_lst}}
}
\concept{eeg_lst}
