% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{eeg_power_band}
\alias{eeg_power_band}
\title{Compute power bands.}
\usage{
eeg_power_band(
  .data,
  .bands = list(delta = c(0.5, 4), theta = c(4, 8), alpha = c(8, 13), beta = c(13, 30)),
  .relative = FALSE,
  ...
)
}
\arguments{
\item{.data}{An \code{eeg_lst} or \code{psd_lst}.}

\item{.bands}{A named list of frequency bands.}

\item{.relative}{Calculate band power relative to total power.}

\item{...}{Other arguments passed to \code{eeg_psd()}}
}
\description{
Compute power bands.
}
\seealso{
Other frequency-based functions: 
\code{\link{eeg_psd}()}
}
\concept{frequency-based functions}
