% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{eeg_psd}
\alias{eeg_psd}
\title{Compute the power spectral density (PSD) of an \code{eeg_lst} object}
\usage{
eeg_psd(
  .data,
  .method = "welch",
  .config = list(window = function(x) 2^ceiling(log2(sqrt(NROW(x)))), overlap = 0.5,
    nfft = NULL, detrend = c("long-mean", "short-mean", "long-linear", "short-linear",
    "none"), range = "half")
)
}
\arguments{
\item{.data}{A \code{eeg_lst} object.}

\item{.method}{"welch" for Welch’s method.}

\item{.config}{See \link[gsignal:pwelch]{gsignal::pwelch}.}
}
\value{
A \code{psd_lst} object
}
\description{
Wrapper of \link[gsignal:pwelch]{gsignal::pwelch} function.
}
\examples{
psd_faces <- eeg_psd(data_faces_ERPs) 
psd_faces \%>\% 
   eeg_select(F3) \%>\%
   plot()     

}
