% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{eeg_rereference}
\alias{eeg_rereference}
\title{Re-reference a channel or group of channels.}
\usage{
eeg_rereference(.data, ..., .ref = NULL, na.rm = FALSE)
}
\arguments{
\item{.data}{An eeg_lst object.}

\item{...}{Channels to include. All the channels by default, but eye channels should be removed.}

\item{.ref}{Channels that will be averaged as the reference. tidyselect compatible.}

\item{na.rm}{a logical evaluating to \code{TRUE} or \code{FALSE}
    indicating whether \code{NA} values should be stripped before the
    computation proceeds.}
}
\value{
An  eeg_lst with some channels re-referenced.
}
\description{
Re-reference a channel or group of channels.
}
\details{
Notice that this function will also rereference the eye electrodes unless excluded. See examples.
}
\examples{
# Re-reference all channels using the left mastoid excluding the eye electrodes.
data_faces_ERPs_M1 <- data_faces_ERPs \%>\%
  eeg_rereference(-EOGV, -EOGH, .ref = M1)

# Re-reference using the linked mastoids excluding the eye electrodes.
data_faces_ERPs_M1M2 <- data_faces_ERPs \%>\%
  eeg_rereference(-EOGV, -EOGH, .ref = c(M1, M2))
}
\seealso{
Other preprocessing functions: 
\code{\link{eeg_baseline}()},
\code{\link{eeg_downsample}()},
\code{\link{eeg_ica}()},
\code{\link{eeg_ica_keep}()},
\code{\link{eeg_segment}()},
\code{\link{filt}}
}
\concept{preprocessing functions}
