% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channel_functions.R
\name{eeg_rereference}
\alias{eeg_rereference}
\title{Re-reference a channel or group of channels.}
\usage{
eeg_rereference(.data, ..., ref = NULL, na.rm = FALSE)
}
\arguments{
\item{.data}{An eeg_lst object.}

\item{...}{Channels to include. All the channels by default, but eye channels channels should be removed.}

\item{ref}{Character vector of channels that will be averaged as the reference.}

\item{na.rm}{a logical value indicating whether \code{NA}
    values should be stripped before the computation proceeds.}
}
\value{
An  eeg_lst with some channels re-referenced.
}
\description{
Re-reference a channel or group of channels.
}
\details{
Notice that this function will update the channels one by one when used inside a mutate and not all at the same time.
}
\examples{
\dontrun{
# Re-reference all channels used the linked mastoids (average of the two mastoids)

faces_segs \%>\% eeg_rereference(ref = c("M1", "M2"))
}
}
\seealso{
Other channel: \code{\link{ch_baseline}},
  \code{\link{channel_dbl}}, \code{\link{chs_mean}},
  \code{\link{is_channel_dbl}}
}
\concept{channel}
