% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_transformation.R
\name{eeg_segment}
\alias{eeg_segment}
\title{Segment an eeg_lst.}
\usage{
eeg_segment(x, ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{...}{Description of the event.}

\item{unit}{Unit}
}
\value{
An \code{eeg_lst}.
}
\description{
Subdivide of the EEG into different segments or epochs. When there is no
segmentation, the \code{eeg_lst} contain one segment. (Fieldtrip calls the
segment "trials".) The limits of \code{segment} are inclusive: If, for
example, \code{lim = c(0,0)}, the segment would contain only sample 1.
}
