% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentation.R
\name{eeg_segment}
\alias{eeg_segment}
\alias{eeg_segment.eeg_lst}
\title{Segment an eeg_lst.}
\usage{
eeg_segment(.data, ...)

\method{eeg_segment}{eeg_lst}(.data, ..., lim = c(-0.5, 0.5), end,
  unit = "s")
}
\arguments{
\item{.data}{An \code{eeg_lst} object.}

\item{...}{Description of the event.}

\item{lim}{Vector indicating the time before and after the event. Or dataframe with two columns, with nrow=total number of segments}

\item{end}{Description of the event that indicates the end of the segment, if this is used, \code{lim} is ignored.}

\item{unit}{Unit}

\item{.recording_col}{Column in the segments table indicating to which .recording or file each segment belongs.}
}
\value{
An \code{eeg_lst}.
}
\description{
Subdivide of the EEG into different segments or epochs. When there is no
segmentation, the \code{eeg_lst} contain one segment. (Fieldtrip calls the
segment "trials".) The limits of \code{segment} are inclusive: If, for
example, \code{lim = c(0,0)}, the segment would contain only sample 1.
}
