% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_ext.R
\name{eeg_slice_signal}
\alias{eeg_slice_signal}
\alias{slice_signal}
\title{Choose samples by the position}
\usage{
eeg_slice_signal(.data, ..., .preserve = FALSE)

slice_signal(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{An eeg_lst object.}

\item{...}{Integer row values.}

\item{.preserve}{Not in use.}
}
\description{
Choose samples by their ordinal position in the signal table. Grouped eeg_lst object use the ordinal position in the signal table within the group.
}
\seealso{
Other tidyverse-like functions: 
\code{\link{dplyr_verbs}},
\code{\link{drop_incomplete_segments}()},
\code{\link{eeg_bind}()}
}
\concept{tidyverse-like functions}
