% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eegble.R
\docType{package}
\name{eegble}
\alias{eegble}
\alias{eegble-package}
\title{eegble: A package for flexible manipulation of EEG data.}
\description{
The eegble package provides functions to load the data, manipulate it and export it to a dataframe.
}
