% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constr.R, R/constructors.R, R/eegble.R
\docType{package}
\name{eegble}
\alias{eegble}
\alias{eegble}
\alias{eegble}
\alias{eegble-package}
\title{Builds an eegble.}
\usage{
eegble(signal, events, segments)

eegble(signal, events, segments)
}
\value{
A valid eggble.
}
\description{
Builds an eegble.

Builds an eegble.

The eegble package provides functions to load the data, manipulate it and export it to a dataframe.
}
