% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R, R/eegble.R
\docType{package}
\name{eegble}
\alias{eegble}
\alias{eegble}
\alias{eegble-package}
\title{Builds an eegble.}
\usage{
eegble(signal, events, segments)
}
\value{
A valid eggble.
}
\description{
Builds an eegble.

The eegble package provides functions that can be used on already pre-processed (i.e.,
filtering, artifact rejects, ICA, etc has already beed done) EEG files (at least for now).
The package mainly provides dplyr-like functions to manipulate the EEG data, and ggplot wrapper functions.
}
