% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{eeguana-deprecated}
\alias{eeguana-deprecated}
\alias{downsample}
\alias{segment}
\alias{interpolate_tbl}
\alias{event_to_ch_NA}
\title{Deprecated functions in eeguana}
\usage{
downsample(x, q = 2, max_sample = NULL, ...)

segment(x, ...)

interpolate_tbl(.data, ...)

event_to_ch_NA(x, ...)
}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item 'downsample()' should be changed to 'eeg_downsample()'
\item 'segment()' should be changed to 'eeg_segment()'
\item 'interpolate_tbl()' should be changed to 'eeg_interpolate_tbl()'
\item 'event_to_ch_NA()' should be changed to 'eeg_intervals_to_NA()'
}
}
