% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{eeguana-deprecated}
\alias{eeguana-deprecated}
\alias{downsample}
\alias{segment}
\alias{interpolate_tbl}
\alias{event_to_ch_NA}
\alias{eeg_intervals_to_NA}
\alias{ch_filt_low_pass}
\alias{ch_filt_high_pass}
\alias{ch_filt_band_pass}
\alias{ch_filt_stop_pass}
\alias{events}
\alias{events<-}
\title{Deprecated functions in eeguana}
\usage{
downsample(x, q = 2, max_sample = NULL, ...)

segment(x, ...)

interpolate_tbl(.data, ...)

event_to_ch_NA(x, ...)

eeg_intervals_to_NA(x, ...)

ch_filt_low_pass(x, ...)

ch_filt_high_pass(x, ...)

ch_filt_band_pass(x, ...)

ch_filt_stop_pass(x, ...)

events(x, ...)

events(x, ...) <- value
}
\arguments{
\item{x}{An eeg_lst object.}

\item{q}{integer factor(s) to downsample by.}

\item{max_sample}{Optionally, the (approximated) maximum sample number can be defined here, which is at least half of the total number of samples}

\item{...}{Not in use.}

\item{.data}{An \code{eeg_lst} object or a long table with amplitudes..}
}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item 'downsample()' should be changed to 'eeg_downsample()'
\item 'segment()' should be changed to 'eeg_segment()'
\item 'interpolate_tbl()' should be changed to 'eeg_interpolate_tbl()'
\item 'event_to_ch_NA()' should be changed to 'eeg_intervals_to_NA()'
}
}
