% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeguana.R
\docType{package}
\name{eeguana-package}
\alias{eeguana}
\alias{eeguana-package}
\title{eeguana: A package for flexible manipulation of EEG data}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

The package *eeguana* provides a framework for doing simple pre-processing with specialized functions (starting with `eeg_` and `ch_`) and manipulating EEG data with *dplyr* verbs (e.g., `mutate`, `filter`, `summarize`) extended to a new class `eeg_lst`, and `ggplot` wrapper functions.  The new class is inspired by tidyverse principles but it's not really "tidy" (due to space considerations), it's a list of (i) a wide table that contains the signal amplitudes at every sample point of the EEG,  (ii) an events table with information about markers (or triggers), blinks and other exported information, and (iii) a long table with experimental information, such as participant (`.recording`), conditions, etc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bnicenboim/eeguana}
  \item Report bugs at \url{https://github.com/bnicenboim/eeguana/issues}
}

}
\author{
\strong{Maintainer}: Bruno Nicenboim \email{bruno.nicenboim@uni-potsdam.de}

Other contributors:
\itemize{
  \item Kate Stone \email{kate.stone@uni-potsdam.de} [contributor]
}

}
\keyword{internal}
