% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eeguana.R
\docType{package}
\name{eeguana-package}
\alias{eeguana}
\alias{eeguana-package}
\title{eeguana: Flexible Manipulation of EEG Data}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Flexible manipulation of EEG data in R. It provides a data.table 
    powered framework for manipulating EEG data with dplyr-based functions
    (e.g., `mutate`, `filter`, `summarize`) extended to a new class, eeg_lst,
    other EEG-specialized functions, and `ggplot` wrapper functions.  The new
    class is inspired by tidyverse principles but it's not really "tidy" (due
    to space considerations), it's a list of (i) a wide data table (signal_tbl)
    that contains the signal amplitudes at every sample point of the EEG,  (ii)
    an events data table with information about markers (or triggers), blinks
    and other exported information, and (iii) a long table with experimental
    information, such as participant number (.recording), conditions, etc.
    eeguana can do only basic pre-processing for now.
}
\seealso{
Useful links:
\itemize{
  \item \url{<https://github.com/bnicenboim/eeguana>}
  \item Report bugs at \url{<https://github.com/bnicenboim/eeguana/issues>}
}

}
\author{
\strong{Maintainer}: Bruno Nicenboim \email{bruno.nicenboim@uni-potsdam.de}

Other contributors:
\itemize{
  \item Kate Stone \email{kate.stone@uni-potsdam.de} [contributor]
}

}
\keyword{internal}
