% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_trans.R
\name{event_to_NA}
\alias{event_to_NA}
\title{Remove (transform to NA) problematic events from an eegble.}
\usage{
event_to_NA(x, ..., all_chans = FALSE, entire_seg = FALSE,
  drop_events = TRUE)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{...}{Description of the problematic event.}

\item{all_chans}{If set to TRUE, 
it will remove samples from all channels (Default:  all_chans = FALSE).}

\item{entire_seg}{If set to TRUE, it will remove the entire segment 
(Default: entire_seg = FALSE).}

\item{drop_events}{}
}
\value{
An eegbl.
}
\description{
Remove (transform to NA) problematic events from an eegble.
}
\examples{

}
