% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_trans.R
\name{event_to_NA}
\alias{event_to_NA}
\title{Remove (transform to NA) problematic events from an eegble.}
\usage{
event_to_NA(x, ..., ind_channel = TRUE, entire_seg = FALSE)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{...}{Description of the problematic event.}
}
\value{
An eegbl.
}
\description{
Remove (transform to NA) problematic events from an eegble.
}
