% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{events_tbl}
\alias{events_tbl}
\alias{events_tbl<-}
\title{Functions to get or set the events table of an eeg_lst object.}
\usage{
events_tbl(.data, ...)

events_tbl(.data) <- value
}
\arguments{
\item{.data}{An eeg_lst object.}

\item{...}{Not in use.}

\item{value}{An events table.}
}
\value{
A table.
}
\description{
Functions to get or set the events table of an eeg_lst object.
}
