% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filt}
\alias{filt}
\alias{eeg_filt_low_pass}
\alias{eeg_filt_high_pass}
\alias{eeg_filt_band_pass}
\alias{eeg_filt_band_stop}
\alias{eeg_filt_low_pass.eeg_lst}
\alias{eeg_filt_high_pass.eeg_lst}
\alias{eeg_filt_band_stop.eeg_lst}
\alias{eeg_filt_band_pass.eeg_lst}
\title{Apply a zero-phase low-pass, high-pass, band-pass, or band-stop filter.}
\usage{
eeg_filt_low_pass(.data, ..., freq = NULL, config = list())

eeg_filt_high_pass(.data, ..., freq = NULL, config = list())

eeg_filt_band_pass(.data, ..., freq = NULL, config = list())

eeg_filt_band_stop(.data, ..., freq = NULL, config = list())

\method{eeg_filt_low_pass}{eeg_lst}(.data, ..., freq = NULL,
  config = list())

\method{eeg_filt_high_pass}{eeg_lst}(.data, ..., freq = NULL,
  config = list())

\method{eeg_filt_band_stop}{eeg_lst}(.data, ..., freq = NULL,
  config = list())

\method{eeg_filt_band_pass}{eeg_lst}(.data, ..., freq = NULL,
  config = list())
}
\arguments{
\item{.data}{A channel or an eeg_lst.}

\item{...}{Channels to apply the filters to. All the channels by default.}

\item{freq}{A single cut frequency for \code{eeg_filt_low_pass} and \code{filt_high_pass_ch}, two edges for  \code{filt_band_pass_ch} and \code{filt_stop_pass_ch}.}

\item{config}{Other parameters passed in a list to the ICA method. (Not implemented)}
}
\value{
A channel or an eeg_lst.
}
\description{
Apply a zero-phase low-pass, high-pass, band-pass, or band-stop filter of the  FIR (finite impulse response) class to every segment of an \code{eeg_lst}. These filters are adapted from the default filters in \href{https://mne-tools.github.io}{MNE package} (v 0.0.17.1)  of \href{https://www.python.org/}{python}. For background information about the FIR vs IIR filters, see (here)\link{https://martinos.org/mne/dev/auto_tutorials/plot_background_filtering.html#sphx-glr-auto-tutorials-plot-background-filtering-py},.
}
\details{
\itemize{
\item \code{eeg_filt_low_pass()} Low-pass or high-cut filter.
\item \code{eeg_filt_high_pass()} High-pass or low-cut filter.
\item \code{eeg_filt_band_pass()} Band-pass filter.
\item \code{eeg_filt_stop_pass()} Stop-pass filter.
}
}
\examples{
data("data_faces_ERPs")
data_ERPs_filtered <- data_faces_ERPs \%>\% eeg_filt_low_pass(freq = 5)
# Compare the ERPs
data_faces_ERPs \%>\% select(O1, O2, P7, P8) \%>\% plot() + facet_wrap(~.source)
data_ERPs_filtered \%>\% select(O1, O2, P7, P8) \%>\% plot() + facet_wrap(~.source)
}
