% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filters.R
\name{filt}
\alias{filt}
\alias{ch_filt_low_pass}
\alias{ch_filt_high_pass}
\alias{ch_filt_band_pass}
\alias{ch_filt_stop_pass}
\alias{ch_filt_high_pass.channel_dbl}
\alias{ch_filt_band_pass.channel_dbl}
\alias{ch_filt_stop_pass.channel_dbl}
\alias{ch_filt_low_pass.eeg_lst}
\alias{ch_filt_high_pass.eeg_lst}
\alias{ch_filt_stop_pass.eeg_lst}
\alias{ch_filt_band_pass.eeg_lst}
\title{Apply a butterworth IIR filter.}
\usage{
ch_filt_low_pass(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

ch_filt_high_pass(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

ch_filt_band_pass(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

ch_filt_stop_pass(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

\method{ch_filt_high_pass}{channel_dbl}(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

\method{ch_filt_band_pass}{channel_dbl}(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

\method{ch_filt_stop_pass}{channel_dbl}(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

\method{ch_filt_low_pass}{eeg_lst}(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

\method{ch_filt_high_pass}{eeg_lst}(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

\method{ch_filt_high_pass}{eeg_lst}(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

\method{ch_filt_stop_pass}{eeg_lst}(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)

\method{ch_filt_band_pass}{eeg_lst}(x, band_edge = NULL, order = 6,
  direction = "twopass", ...)
}
\arguments{
\item{x}{A channel or an eeg_lst.}

\item{band_edge}{A single cut frequency for \code{filt_low_pass_ch} and \code{filt_high_pass_ch}, two edges for  \code{filt_band_pass_ch} and \code{filt_stop_pass_ch}.}

\item{order}{Filter order (default is 6).}

\item{direction}{"twopass" (default) filters the signal twice (once forwards, then again backwards).}

\item{...}{The sample rate can be included as \code{sampling_rate}, when the function is called outside mutate/summarize.}
}
\value{
A channel or an eeg_lst.
}
\description{
Apply a Butterworth IIR filter using \code{signal::filt_filt}. Based on
Matt Craddock's code of \code{eegUtils} \email{matt@mattcraddock.com}.
}
\details{
\itemize{
\item \code{filt_low_pass_ch()} Low-pass or high-cut filter.
\item \code{filt_high_pass_ch()} High-pass or low-cut filter.
\item \code{filt_band_pass_ch()} Band-pass filter.
\item \code{filt_stop_pass_ch()} Stop-pass filter.
}
}
\examples{
\dontrun{

faces_segs \%>\% filt_low_pass_ch(band_edge = 100, order = 4)
}
}
