% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{info}
\alias{info}
\alias{channel_names}
\alias{nchannels}
\alias{channels_tbl}
\alias{sampling_rate}
\alias{duration}
\alias{nsamples}
\title{Display information of the eegble object.}
\usage{
channel_names(x, ...)

nchannels(x, ...)

channels_tbl(x, ...)

sampling_rate(x, ...)

duration(x, ...)

nsamples(x, ...)
}
\arguments{
\item{x}{An eegble object.}
}
\description{
\itemize{
\item \code{nchannels()}: Returns the number of channels.
\item \code{chan_names()}: Returns a vector with the name of the channels.
\item \code{channels()}: Returns a data frame (tibble) with information about the channels.
\item \code{info()}: Returns a data frame (tibble) with information about the EEG recording.
\item \code{sampling_rate()}: Returns the sampling rate.
\item \code{reference()}: Returns the reference.
\item \code{duration()}: Returns the duration of the recording (or segments).
\item \code{nsamples()}: Returns the number of samples of the recording (or segments).
}
}
