% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{info}
\alias{info}
\alias{nchan}
\alias{chan_names}
\alias{chan_info}
\alias{gral_info}
\alias{srate}
\alias{duration}
\title{Display information of the eegble object.}
\usage{
nchan(x)

chan_names(x)

chan_info(x)

gral_info(x)

srate(x)

duration(x)
}
\arguments{
\item{x}{An eegble object.}
}
\description{
\itemize{
\item \code{nchan()}: Returns the number of channels.
\item \code{chan_names()}: Returns a vector with the name of the channels.
\item \code{chan_info()}: Returns a dataframe (tibble) with information about the channels.
\item \code{gral_info()}: Returns a dataframe (tibble) with information about the EEG recording.
\item \code{srate()}: Returns the sampling rate.
\item \code{duration()}: Returns the duration of the recording (or segments).
}
}
