% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{info}
\alias{info}
\alias{channel_names}
\alias{nchannels}
\alias{nsamples}
\title{Display information of the eegble object.}
\usage{
channel_names(x, ...)

nchannels(x, ...)

nsamples(x, ...)
}
\arguments{
\item{x}{An eegble object.}
}
\description{
\itemize{
\item \code{nchannels()}: Returns the number of channels.
\item \code{channel_names()}: Returns a vector with the name of the channels.
\item \code{nchannels()}: Returns the number of channels.
\item \code{nsamples()}: Returns the number of samples of the recording (or segments).
\item \code{summary()}: Prints a summary of the eegble object. (It also returns a list.)
\item \code{count_complete_cases_tbl()}: Returns a table with the number of complete segments by specified groups.
}
}
