% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{install_py_eeguana}
\alias{install_py_eeguana}
\title{Install the Python packages needed for \code{eeguana}}
\usage{
install_py_eeguana(method = c("auto", "virtualenv", "conda"),
                     conda = "auto",
                     version = "default",
                     envname = "r-eeguana",
                     restart_session = TRUE,
                     conda_python_version = NULL,
                     ...,
                     pip_ignore_installed = FALSE,
                     new_env = identical(envname, "r-eeguana"),
                     python_version = NULL)
}
\arguments{
\item{conda}{Specifies the conda binary to use. Default is 'auto'.}

\item{version}{The Python version to use. Default is 'default', automatically selected.}

\item{envname}{Name of the virtual environment. Default is 'r-eeguana'.}

\item{restart_session}{Logical, whether to restart the R session after installation.
Default is TRUE.}

\item{conda_python_version}{Python version for conda environments.}

\item{...}{Additional arguments passed to \code{reticulate::py_install}.}

\item{pip_ignore_installed}{Logical, whether to ignore already installed packages.
Default is FALSE.}

\item{new_env}{Logical, whether to create a new environment if \code{envname} is 'r-eeguana'.
Default is the identity of \code{envname}.}

\item{python_version}{Specifies the Python version for the environment.}

\item{method}{A character vector specifying the environment management method.
Options are 'auto', 'virtualenv', and 'conda'. Default is 'auto'.}
}
\value{
The function returns \code{NULL} invisibly, but outputs a message on successful installation.
}
\description{
\code{install_py_eeguana} function facilitates the installation of Python packages needed for using \code{eeguana} within an R environment,
utilizing the \code{reticulate} package for managing Python environments. It supports various installation methods,
environment settings, and Python versions.
}
\details{
This function automatically selects the appropriate method for environment management and Python installation,
with a focus on virtual and conda environments. It ensures flexibility in dependency management and Python version control.
If a new environment is created, existing environments with the same name are removed.
}
