% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{is_channel_dbl}
\alias{is_channel_dbl}
\title{Test if the object is a channel
This function returns  TRUE for channels.}
\usage{
is_channel_dbl(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{sampl} class.
}
\description{
Test if the object is a channel
This function returns  TRUE for channels.
}
\seealso{
Other channel: \code{\link{ch_baseline}},
  \code{\link{channel_dbl}}, \code{\link{chs_mean}},
  \code{\link{eeg_rereference}}
}
\concept{channel}
