% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{is_channel_dbl}
\alias{is_channel_dbl}
\title{Test if the object is a channel or EOG channel}
\usage{
is_channel_dbl(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{channel_dbl} class.
}
\description{
\itemize{
\item \code{is_channel_dbl()} returns TRUE for all  channels including EOG channels.
\item \code{is_eog_channel_dbl()} returns TRUE only for EOG channels.
}
}
\seealso{
Other channel: 
\code{\link{as_channel_dbl}()},
\code{\link{channel_dbl}()}
}
\concept{channel}
