% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{is_psd_lst}
\alias{is_psd_lst}
\title{Test if the object is a psd_lst.}
\usage{
is_psd_lst(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{psd_lst} class.
}
\description{
This function returns  TRUE for psd_lsts.
}
\seealso{
Other psd_lst: 
\code{\link{psd_lst}()}
}
\concept{psd_lst}
