% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{is_sample_int}
\alias{is_sample_int}
\title{Test if the object is a sample
This function returns  TRUE for samples.}
\usage{
is_sample_int(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{sample} class.
}
\description{
Test if the object is a sample
This function returns  TRUE for samples.
}
\seealso{
Other sample_int: 
\code{\link{sample_int}()}
}
\concept{sample_int}
