% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signal_tbl.R
\name{is_signal_tbl}
\alias{is_signal_tbl}
\title{Test if the object is a  signal_tbl}
\usage{
is_signal_tbl(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if the object inherits from the \code{signal_tbl} class.
}
\description{
This function returns  TRUE for signals.
}
\concept{signal_tbl}
