% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eeg_lst}
\alias{plot.eeg_lst}
\title{Simple plot an eeg_lst object.}
\usage{
\method{plot}{eeg_lst}(x, max_sample = 64000, ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{max_sample}{Downsample to approximately 64000 samples by default.}

\item{...}{Not in use.}
}
\value{
A ggplot object
}
\description{
Simple plot an eeg_lst object.
}
\seealso{
Other plot: \code{\link{plot_gg}}, \code{\link{plot_topo}}
}
\concept{plot}
