% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eegbl}
\alias{plot.eegbl}
\title{Simple plot an eegble object.}
\usage{
\method{plot}{eegbl}(x, max_sample = 2000)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{max_sample}{Downsample to approximately 2000 samples by default.}
}
\value{
A ggplot object
}
\description{
Simple plot an eegble object.
}
