% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_components}
\alias{plot_components}
\title{Generates topographic plots of the components after running ICA on an eeg_lst}
\usage{
plot_components(data, ..., .projection = "polar", .standardize = TRUE)
}
\arguments{
\item{data}{An eeg_ica_lst}

\item{...}{arguments passed to interpolate.}

\item{.projection}{.Projection type for converting the 3D coordinates of the electrodes into 2d coordinates.  .Projection types available: "polar" (default), "orthographic", or  "stereographic"}

\item{.standardize}{Whether to standardize the color scale of each topographic plot.}
}
\description{
Note that unlike \link{plot_topo}, there is no need for faceting, or adding layers.
}
\examples{
# For demonstration only, since ICA won't converge
library(ggplot2)
# Suppressing an important warning:
suppressWarnings(data_faces_10_trials \%>\%
  eeg_ica(-EOGH, -EOGV, -M1, -M2, .method = fast_ICA, .config = list(maxit = 10))) \%>\%
  eeg_ica_keep(ICA1, ICA2) \%>\%
  plot_components() +
  annotate_head() +
  geom_contour() +
  annotate_electrodes(color = "black") +
  theme(legend.position = "none")
}
\seealso{
Other plotting functions: 
\code{\link{annotate_electrodes}()},
\code{\link{annotate_events}()},
\code{\link{annotate_head}()},
\code{\link{eeg_downsample}()},
\code{\link{ggplot.eeg_lst}()},
\code{\link{plot.eeg_lst}()},
\code{\link{plot_in_layout}()},
\code{\link{plot_topo}()},
\code{\link{theme_eeguana}()}

Other ICA functions: 
\code{\link{eeg_ica}()},
\code{\link{eeg_ica_cor_tbl}()},
\code{\link{eeg_ica_keep}()},
\code{\link{eeg_ica_show}()},
\code{\link{eeg_ica_summary_tbl}()},
\code{\link{eeg_ica_var_tbl}()}

Other topographic plots and layouts: 
\code{\link{layout_32_1020}},
\code{\link{plot_in_layout}()},
\code{\link{plot_topo}()}
}
\concept{ICA functions}
\concept{plotting functions}
\concept{topographic plots and layouts}
