% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_components}
\alias{plot_components}
\alias{plot_components.eeg_ica_lst}
\alias{plot_ica.eeg_ica_lst}
\title{Generates topographic plots of the components after running ICA on an eeg_lst}
\usage{
plot_components(data, ...)

\method{plot_components}{eeg_ica_lst}(data, ..., projection = "polar",
  standardize = TRUE)

\method{plot_ica}{eeg_ica_lst}(data, ..., eog = list(...),
  .recording = NULL, samples = 1:300, order = c("cor", "ica"),
  max_sample = 2400, topo_config = list(projection = "polar",
  standardize = TRUE), interp_config = list(...))
}
\arguments{
\item{data}{An eeg_ica_lst}

\item{...}{If data are an \code{eeg_lst} or \code{ica_lst}, these are arguments passed to \code{eeg_interpolate_tbl}, such as, radius, size, etc.}

\item{projection}{Projection type for converting the 3D coordinates of the electrodes into 2d coordinates. Projection types available: "polar" (default), "orthographic", or  "stereographic"}
}
\description{
Note that unlike \link{plot_topo}, there is no need for faceting, or adding layers.
}
\seealso{
Other plotting functions: \code{\link{annotate_events}},
  \code{\link{annotate_head}},
  \code{\link{eeg_downsample}},
  \code{\link{ggplot.eeg_lst}}, \code{\link{plot.eeg_lst}},
  \code{\link{plot_in_layout}}, \code{\link{plot_topo}},
  \code{\link{theme_eeguana}}

Other ICA functions: \code{\link{eeg_ica_cor_tbl}},
  \code{\link{eeg_ica_keep}}, \code{\link{eeg_ica_show}},
  \code{\link{eeg_ica}}
}
\concept{ICA functions}
\concept{plotting functions}
