% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gg}
\alias{plot_gg}
\title{ggplot object based on an eegble object.}
\usage{
plot_gg(x, ...)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{max_sample}{Downsample to approximately 2000 samples by default.}
}
\value{
A ggplot object
}
\description{
ggplot object based on an eegble object.
}
