% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gg}
\alias{plot_gg}
\title{ggplot object based on an eeg_lst object.}
\usage{
plot_gg(.data, ...)
}
\arguments{
\item{x}{An \code{eeg_lst} object.}

\item{max_sample}{Downsample to approximately 2000 samples by default.}
}
\value{
A ggplot object
}
\description{
ggplot object based on an eeg_lst object.
}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_topo}}
}
\concept{plot}
