% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{plot_gg}
\alias{plot_gg}
\title{ggplot object based on an eegble object.}
\usage{
plot_gg(x, ..., thinning = "auto")
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{thinning}{Automatic by default, but integers can be used.}
}
\value{
A ggplot object
}
\description{
ggplot object based on an eegble object.
}
