% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_in_layout}
\alias{plot_in_layout}
\alias{plot_in_layout.gg}
\title{Place channels in a layout.}
\usage{
plot_in_layout(plot, ...)

\method{plot_in_layout}{gg}(plot, projection = "polar", ratio = c(1,
  1), ...)
}
\arguments{
\item{plot}{A ggplot object with channels}

\item{...}{Not in use.}

\item{projection}{Projection type for converting the 3D coordinates of the electrodes into 2d coordinates. Projection types available: "polar" (default), "orthographic", or  "stereographic"}

\item{ratio}{Ratio of the individual panels}
}
\value{
A ggplot object
}
\description{
Reorganizes a ggplot so that the plot of each channel is in their correct position in the scalp.
}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_gg}}, \code{\link{plot_topo}}
}
\concept{plot}
