% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_in_layout}
\alias{plot_in_layout}
\alias{plot_in_layout.gg}
\title{Place channels in a layout.}
\usage{
plot_in_layout(plot, ...)

\method{plot_in_layout}{gg}(plot, projection = "polar", size = c(1, 1),
  ...)
}
\arguments{
\item{plot}{A ggplot object with channels}

\item{...}{}

\item{projection}{"polar" (default), "orthographic", or  "stereographic"}

\item{size}{}

\item{...}{}
}
\value{
A ggplot object
}
\description{
Reorganizes a ggplot so that the plot of each channel is in their correct position in the scalp.
}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_gg}}, \code{\link{plot_topo}}
}
\concept{plot}
