% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_in_layout}
\alias{plot_in_layout}
\alias{plot_in_layout.gg}
\title{Arrange ERP plots according to scalp layout}
\usage{
plot_in_layout(plot, ...)

\method{plot_in_layout}{gg}(plot, projection = "polar", ratio = c(1,
  1), ...)
}
\arguments{
\item{plot}{A ggplot object with channels}

\item{...}{Not in use.}

\item{projection}{Projection type for converting the 3D coordinates of the electrodes into 2d coordinates. Projection types available: "polar" (default), "orthographic", or  "stereographic"}

\item{ratio}{Ratio of the individual panels}
}
\value{
A ggplot object
}
\description{
Arranges a ggplot so that the facet for each channel appears in its position
on the scalp.
}
\details{
This function requires two steps: first, a ggplot object must be created with
ERPs facetted by channel (\code{.key}).
Then, the ggplot object is called in \code{plot_in_layout}. The function uses grobs
arranged according to .x .y coordinates extracted from the \code{eeg_lst} object, by
default in polar arrangement. The arrangement can be changed with the \code{projection}
argument. White space in the plot can be reduced by changing \code{ratio}.

Additional components such as titles and annotations should be added to the
plot object using \code{+} exactly as you would for \code{ggplot::ggplot}.
Title and legend adjustments will be treated as applying to the
whole plot object, while other theme adjustments will be treated as applying
to individual facets. x-axis and y-axis labels cannot be added at this stage.
}
\examples{

# Create a ggplot object with some grand averaged ERPs
ERP_plot <- data_faces_ERPs \%>\% 
   # select a few electrodes
   select(Fz, FC1, FC2, C3, Cz, C4, CP1, CP2, Pz) \%>\%
   # group by time point and condition
   group_by(.sample, condition) \%>\%
   # compute averages
   summarize_all_ch(mean,na.rm=TRUE) \%>\%
   plot_gg() + 
       # plot the averaged waveforms
       geom_line(aes(color = condition)) +
       # facet by channel
       facet_wrap(~ .key) +  
       # add a legend and title
       theme(legend.position = "bottom") + 
       ggtitle("ERPs for faces vs non-faces") 

# Call the ggplot object with the layout function
plot_in_layout(ERP_plot)
}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_gg}}, \code{\link{plot_topo}}
}
\concept{plot}
