% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_topo}
\alias{plot_topo}
\alias{plot_topo.tbl_df}
\alias{plot_topo.eeg_lst}
\alias{plot_topo.mixing_tbl}
\alias{plot_topo.ica_lst}
\title{A topographic plot.}
\usage{
plot_topo(data, ...)

\method{plot_topo}{tbl_df}(data, value = .value, label = .source, ...)

\method{plot_topo}{eeg_lst}(data, size = 1.2, projection = "polar",
  ...)

\method{plot_topo}{mixing_tbl}(data, size = 1.2, projection = "polar",
  ...)

\method{plot_topo}{ica_lst}(data, size = 1.2, projection = "polar",
  ...)
}
\arguments{
\item{data}{A table of interpolated electrodes as produced by \link{eeg_interpolate_tbl}, or an eeg_lst appropiately grouped.}

\item{...}{Not in use.}

\item{value}{Values used for the interpolation, generally \code{.value} (default).}

\item{label}{Label of the points that are used for the interpolation, generally \code{.source} (default).}

\item{size}{Indicates how much to extrapolate, \code{1} indicates no extrapolation beyond the position of the electrodes.}

\item{projection}{Projection type for converting the 3D coordinates of the electrodes into 2d coordinates. Projection types available: "polar" (default), "orthographic", or  "stereographic"}
}
\value{
A ggplot object
}
\description{
Create a default topographic plot based on an interpolation table.
}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_gg}}, \code{\link{plot_in_layout}}
}
\concept{plot}
