% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_topo}
\alias{plot_topo}
\alias{plot_topo.tbl_df}
\alias{plot_topo.eeg_lst}
\title{Create a topographic plot}
\alias{plot_topo.mixing_tbl}
\alias{plot_topo.ica_lst}
\usage{
plot_topo(data, ...)

\method{plot_topo}{tbl_df}(data, value = .value, label = .source)

\method{plot_topo}{eeg_lst}(data, size = 1.2, projection = "polar",
  ...)

\method{plot_topo}{mixing_tbl}(data, size = 1.2, projection = "polar",
  ...)

\method{plot_topo}{ica_lst}(data, size = 1.2, projection = "polar",
  ...)
}
\arguments{
\item{data}{A table of interpolated electrodes as produced by \link{eeg_interpolate_tbl}, or an eeg_lst appropiately grouped.}

\item{...}{Not in use.}

\item{value}{Values used for the interpolation, generally \code{.value} (default).}

\item{label}{Label of the points that are used for the interpolation, generally \code{.source} (default).}

\item{size}{Indicates how much to extrapolate, \code{1} indicates no extrapolation beyond the position of the electrodes.}

\item{projection}{Projection type for converting the 3D coordinates of the electrodes into 2d coordinates. Projection types available: "polar" (default), "orthographic", or  "stereographic"}
}
\value{
A ggplot object
}
\description{
\code{plot_topo} initializes a ggplot object which takes an eeg_lst object
as its input data. Layers can then be added in the same way as for a
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\details{
Before calling \code{plot_topo}, the eeg_lst object must be appropriately
grouped (e.g. by condition) and/or
summarized into mean values such that each .x .y coordinate has only one
amplitude value. These values can then be interpolated using
\code{\link[=eeg_interpolate_tbl]{eeg_interpolate_tbl()}}, which will display the electrodes
stereographically, or plotted as-is, which will use the default polar display.
\code{plot_topo} called alone
without any further layers will create an unannotated topographical plot.
To add a head and nose, add the layer \code{\link[=annotate_head]{annotate_head()}}. Add
contour lines with \code{\link[ggplot2:geom_contour]{ggplot2::geom_contour()}} and electrode labels
with \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}. These arguments are deliberately not
built into the function so as to allow flexibility in choosing colour, font
size, and even head size, etc.
}
\examples{
```
# Calculate mean amplitude between 100-200 ms and plot the topography
data_faces_ERPs \%>\% 
    # select the time window of interest
    filter(between(as_time(.sample_id, unit = "milliseconds"),100,200)) \%>\% 
    # compute mean amplitude per condition
    group_by(condition) \%>\%
    summarize_all_ch(mean, na.rm = TRUE) \%>\%
    plot_topo() +
        # add a head and nose shape
        annotate_head() + 
        # add contour lines
        geom_contour() +
        # add electrode labels
        geom_text(colour = "black") +
        facet_grid(~condition)

# The same but with interpolation
data_faces_ERPs \%>\% 
    filter(between(as_time(.sample_id, unit = "milliseconds"),100,200)) \%>\% 
    group_by(condition) \%>\%
    summarize_all_ch(mean, na.rm = TRUE) \%>\%
    eeg_interpolate_tbl() \%>\%
    plot_topo() +
        annotate_head() + 
        geom_contour() +
        geom_text(colour = "black") +
        facet_grid(~condition)
```

}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_gg}}, \code{\link{plot_in_layout}}
}
\concept{plot}
