% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_topo}
\alias{plot_topo}
\alias{plot_topo.tbl_df}
\title{A topographic plot.}
\usage{
plot_topo(data, ...)

\method{plot_topo}{tbl_df}(data, x = .x, y = .y, value = amplitude,
  label = channel, ...)
}
\arguments{
\item{data}{A table of interpolated electrodes as produced by \link{interpolate_tbl}}

\item{...}{Others.}

\item{x}{x}

\item{y}{y}

\item{value}{value}

\item{label}{label, generally channel}
}
\value{
A ggplot object
}
\description{
Create a default topographic plot based on an interpolation table.
}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_gg}}, \code{\link{plot_in_layout}}
}
\concept{plot}
