% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_topo}
\alias{plot_topo}
\title{A topographic plot of an eeg_lst object.}
\usage{
plot_topo(data, ...)
}
\arguments{
\item{...}{Various arguments passed to the interpolation method.}

\item{.data}{A table of interpolated electrodes as produced by \link{interpolate_tbl}}

\item{x}{Coordinate x}

\item{y}{Coordinate y}

\item{value}{amplitude (default)}

\item{label}{channel (default)}
}
\value{
A ggplot object
}
\description{
Create a default topographic plot based on the segments of the \code{eeg_lst} object.
}
\details{
The following methods of interpolation are available :
\itemize{
\item \code{"MBA"} (Default) Multilevel B-splines using the function \code{mba.surf}; requires the package \code{MBA}.
}
}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_gg}}
}
\concept{plot}
