% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_topo}
\alias{plot_topo}
\alias{plot_topo.tbl_df}
\alias{plot_topo.eeg_lst}
\title{A topographic plot.}
\usage{
plot_topo(data, ...)

\method{plot_topo}{tbl_df}(data, value = amplitude, label = channel,
  ...)

\method{plot_topo}{eeg_lst}(data, size = 1.2, value = amplitude,
  label = channel, projection = "polar", ...)
}
\arguments{
\item{data}{A table of interpolated electrodes as produced by \link{eeg_interpolate_tbl}, or an eeg_lst appropiately grouped.}

\item{...}{Others.}

\item{value}{value}

\item{label}{label, generally channel}

\item{projection}{"polar" (default), "orthographic", or  "stereographic"}

\item{...}{passed to the interpolation method}
}
\value{
A ggplot object
}
\description{
Create a default topographic plot based on an interpolation table.
}
\seealso{
Other plot: \code{\link{plot.eeg_lst}},
  \code{\link{plot_gg}}, \code{\link{plot_in_layout}}
}
\concept{plot}
