% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{psd_lst}
\alias{psd_lst}
\title{Creates a \code{psd_lst}.}
\usage{
psd_lst(psd_tbl = NULL, segments_tbl = NULL, channels_tbl = NULL)
}
\arguments{
\item{psd_tbl}{See \code{\link[=psd_tbl]{psd_tbl()}}.}

\item{segments_tbl}{A data table of segment numbers and related information. See \code{\link[=segments_tbl]{segments_tbl()}}.}

\item{channels_tbl}{Optionally a table with channels information. See \code{\link[=channels_tbl]{channels_tbl()}}.}
}
\value{
A valid psd_lst.
}
\description{
Creates a \code{psd_lst}.
}
\seealso{
Other psd_lst: 
\code{\link{is_psd_lst}()}
}
\concept{psd_lst}
