% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_dat}
\alias{read_dat}
\title{Helper function to read the dat files directly}
\usage{
read_dat(file, common_info = NULL, chan_info = NULL, events = NULL,
  .id = file, ignore_segs, sep, zero, verbose = verbose)
}
\description{
Helper function to read the dat files directly
}
