% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_edf}
\alias{read_edf}
\title{Read a edf/edf+/bdf file into an eeg_lst object.}
\usage{
read_edf(file, recording = file)
}
\arguments{
\item{file}{A edf/bdf file}

\item{recording}{Recording name (file name, by default). If set to NULL or NA, the patient name will be used.}
}
\value{
An \code{eeg_lst} object.
}
\description{
Read a edf/edf+/bdf file into an eeg_lst object.
}
\seealso{
Other read: \code{\link{read_ft}}, \code{\link{read_vhdr}}
}
\concept{read}
