% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_edf}
\alias{read_edf}
\title{Read an edf/edf+/bdf file into R}
\usage{
read_edf(file, recording = file)
}
\arguments{
\item{file}{A edf/bdf file}

\item{recording}{Recording name (file name, by default). If set to NULL or NA, the patient name will be used.}
}
\value{
An \code{eeg_lst} object.
}
\description{
Creates an eeg_lst object from edf, edf+, and bdf file export formats.
}
\examples{
\dontrun{s1 <- read_edf("./faces.edf", recording = 1)}

}
\seealso{
Other read: \code{\link{read_ft}}, \code{\link{read_vhdr}}
}
\concept{read}
