% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_fif}
\alias{read_fif}
\title{Read a FIF file into R}
\usage{
read_fif(file, .recording = file)
}
\arguments{
\item{file}{A .fif file containing EEG/MEG data.}

\item{.recording}{Recording name, by default is the file name. This parameter is currently not used in the function but reserved for future use or metadata purposes.}
}
\value{
An \code{eeg_lst} object with signal_tbl and event information from a .fif file.
}
\description{
Creates an eeg_lst object from EEG/MEG data in FIF format. The function reads a .fif file using the \code{mne} Python package via \code{reticulate}.
If you do not already have \code{mne} installed in your Python environment, ensure to install it prior to using this function using \link{install_py_eeguana}.
The .fif file is a standard format used by MNE and other neurophysiological data analysis software.
}
\examples{
\dontrun{
eeg_data <- read_fiff("./subject1_raw.fif")
}

}
\seealso{
Other reading and writing functions: 
\code{\link{read_edf}()},
\code{\link{read_ft}()},
\code{\link{read_set}()},
\code{\link{read_vhdr}()}
}
\concept{reading and writing functions}
