% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ft}
\alias{read_ft}
\title{Read a Fieldtrip file into R}
\usage{
read_ft(file, .layout = NULL, .recording = file)
}
\arguments{
\item{file}{A .mat file containing a fieldtrip struct.}

\item{.layout}{A .mat \href{http://www.fieldtriptoolbox.org/template/layout}{layout from Fieldtrip}}

\item{.recording}{Recording name, by default is the file name.}
}
\value{
An \code{eeg_lst} object with signal_tbl and event from a Matlab file.
}
\description{
Creates an eeg_lst object from Matlab exported files.The function reads a .mat
file using \code{R.matlab}. If you do not already have
\code{R.matlab} installed in R, you will need to install it yourself. The .mat file
should have the structure described in
this \href{http://www.fieldtriptoolbox.org/reference/ft_datatype_raw}{Fieldtrip reference article}.
}
\examples{
\dontrun{
s1 <- read_ft("./subject1.mat", layout = "easycapM25.mat", .recording = 1)
}

}
\seealso{
Other reading and writing functions: 
\code{\link{read_edf}()},
\code{\link{read_fif}()},
\code{\link{read_set}()},
\code{\link{read_vhdr}()}
}
\concept{reading and writing functions}
