% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_ft}
\alias{read_ft}
\title{Read a Fieldtrip file into an eeg_lst object (requires R.matlab).}
\usage{
read_ft(file, layout = NULL, recording = file)
}
\arguments{
\item{file}{A .mat file containing a fieldtrip struct.}

\item{layout}{A .mat \href{http://www.fieldtriptoolbox.org/template/layout}{layout from Fieldtrip}}

\item{recording}{Recording name, by default is the file name.}
}
\value{
An \code{eeg_lst} object with signal_tbl and event from a matlab file.
}
\description{
Read a Fieldtrip file into an eeg_lst object (requires R.matlab).
}
\seealso{
Other read: \code{\link{read_vhdr}}
}
\concept{read}
