% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_set}
\alias{read_set}
\title{Read EEGlab set files (Matlab files) into R}
\usage{
read_set(file, .recording = file)
}
\arguments{
\item{file}{A .mat or .set file containing a fieldtrip struct.}

\item{.recording}{Recording name, by default is the file name.}
}
\value{
An \code{eeg_lst} object with signal_tbl and event from a Matlab file.
}
\description{
Creates an eeg_lst object from Matlab exported files. The function reads a .mat or .set file using \code{R.matlab}. If you do not already have \code{R.matlab} installed in R, you will need to install it yourself. The  file should have the structure described in this \href{https://sccn.ucsd.edu/wiki/A05:_Data_Structures}{Data structure article}. This function is experimental (there are many different formats of eeglab files). If your file cannot be opened please open an issue with a link to the file in \href{https://github.com/bnicenboim/eeguana/issues}{github}.
}
\examples{
\dontrun{
s1 <- read_ft("./subject1.set", .recording = 1)
}

}
\seealso{
Other reading functions: 
\code{\link{read_edf}()},
\code{\link{read_ft}()},
\code{\link{read_vhdr}()}
}
\concept{reading functions}
