% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_vhdr}
\alias{read_vhdr}
\title{Read a BrainVision file into an eegble object.}
\usage{
read_vhdr(file, ignore_segs = FALSE, sep = type == "New Segment",
  zero = type == "Time 0", verbose = TRUE)
}
\arguments{
\item{file}{A vhdr file in a folder that contains a .vmrk and .dat files}

\item{ignore_segs}{Ignore BrainVision segmentation. By default, FALSE.}

\item{sep}{Segment separation marker. By default: type == "New Segment"}

\item{zero}{Time zero marker. By default: type == "Time 0"}

\item{verbose}{Prints message with information.}
}
\value{
An \code{eegble} object with signals and event from file.dat, file.vhdr, and file.vmrk.
}
\description{
Read a BrainVision file into an eegble object.
}
