% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_vhdr}
\alias{read_vhdr}
\title{Read a BrainVision file into an eegble object.}
\usage{
read_vhdr(file, sep = type == "New Segment", zero = type == "Time 0",
  recording = file)
}
\arguments{
\item{file}{A vhdr file in a folder that contains a .vmrk and .dat files}

\item{sep}{Segment separation marker. By default: type == "New Segment"}

\item{zero}{Time zero marker. By default: type == "Time 0"}

\item{recording}{Recording name, by default is the file name.}

\item{ignore_segs}{Ignore BrainVision segmentation. By default, FALSE.}
}
\value{
An \code{eegble} object with signal and event from file_name.dat, 
file_name.vhdr, and file_name.vmrk.
}
\description{
Read a BrainVision file into an eegble object.
}
