% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_trans.R
\name{segment}
\alias{segment}
\title{Segments an eegble.}
\usage{
segment(x, ..., lim = c(-0.5, 0.5), unit = "seconds")
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{...}{Description of the event.}

\item{lim}{Vector indicating the time before and after the event. Or matrix with two columns, with nrow=total number of segments}
}
\value{
An eegbl.
}
\description{
EXPLAIN WHAT A SEGMENT IS/ DIFFERENCE WITH EPOCH (LINK?) + EXAMPLES OF SEGMENT WITH +Inf.
Fieldtrip calls the segment "trials". The limits are inclusive, the event is
 set as happening in the sample 1. If, for example, lim =c(0,0), the segment would contain only one sample.
}
