% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables_transformation.R
\name{segment}
\alias{segment}
\title{Segments an eegble.}
\usage{
segment(x, ...)
}
\arguments{
\item{x}{An \code{eegble} object.}

\item{...}{Description of the event.}

\item{lim}{Vector indicating the time before and after the event. Or matrix with two columns, with nrow=total number of segments}
}
\value{
An \code{eegble}.
}
\description{
Subdivides of the EEG into different segments or epochs. When there is no
segmentation, the \code{eegble} contain one segment. (Fieldtrip calls the
segment "trials".) The limits of \code{segment} are inclusive: If, for
example, lim =c(0,0), the segment would contain only sample 1.
}
