% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{signal_tbl}
\alias{signal_tbl}
\title{Builds a signal_tbl table.}
\usage{
signal_tbl(signal_matrix, ids, sample_ids, channel_info)
}
\arguments{
\item{signal_matrix}{}

\item{ids}{}

\item{sample_ids}{}

\item{channel_info}{}
}
\value{
A valid signal_tbl table.
}
\description{
Builds a signal_tbl table.
}
\seealso{
Other signal_tbl: \code{\link{is_signal_tbl}}
}
\concept{signal_tbl}
