% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{signal_tbl}
\alias{signal_tbl}
\title{Builds a signal_tbl table}
\usage{
signal_tbl(signal_matrix = NULL, ids = NULL, sample_ids = NULL,
  channel_info = NULL)
}
\arguments{
\item{signal_matrix}{Matrix or table of channels with their signal.}

\item{ids}{Integers indicating to which group the row of the signal matrix belongs.}

\item{sample_ids}{Vector of integers.}

\item{channel_info}{A table with information about each channel (such as the one produced by `channels_tbl``)}
}
\value{
A valid signal_tbl table.
}
\description{
The eeg_lst \code{signal} table is organised into columns representing timestamps
(\code{.sample_id}) and individual electrodes. Each \code{.sample_id} corresponds to
1 sample in the original recording, i.e. if the sampling rate of the EEG
recording is 500 Hz, then each \code{.sample_id} corresponds to 2 milliseconds.
These timestamps correspond to \code{.sample_0} in the \code{events} table, which
displays only the timestamps where logged events began.
}
\seealso{
Other signal_tbl: \code{\link{is_signal_tbl}}
}
\concept{signal_tbl}
