% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{signal_tbl}
\alias{signal_tbl}
\title{Builds a signal_tbl table.}
\usage{
signal_tbl(signal_matrix = NULL, ids = NULL, sample_ids = NULL,
  channel_info = NULL)
}
\arguments{
\item{signal_matrix}{Matrix or table of channels with their signal.}

\item{ids}{Integers indicating to which group the row of the signal matrix belongs.}

\item{sample_ids}{Vector of integers.}

\item{channel_info}{A table with information about each channel (such as the one produced by `channels_tbl``)}
}
\value{
A valid signal_tbl table.
}
\description{
Builds a signal_tbl table.
}
\seealso{
Other signal_tbl: \code{\link{is_signal_tbl}}
}
\concept{signal_tbl}
