% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl.R
\name{signal_tbl}
\alias{signal_tbl}
\title{Function to get the signal table of an eeg_lst object.}
\usage{
signal_tbl(.data, ...)
}
\arguments{
\item{.data}{An eeg_lst object.}

\item{...}{Not in use.}
}
\value{
A table.
}
\description{
Function to get the signal table of an eeg_lst object.
}
