% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{summaries}
\alias{summaries}
\alias{channel_names}
\alias{channel_ica_names}
\alias{nchannels}
\alias{component_names}
\alias{ncomponents}
\alias{nsamples}
\title{Display information of the eeg_lst object.
\code{*_names()} functions return a vector of names, \code{n*()} return the number of
elements channels or components. Components are only available after running \code{\link[=eeg_ica]{eeg_ica()}}. \code{channel_ica_names()} refers to channels used in the ICA.}
\usage{
channel_names(x, ...)

channel_ica_names(x, ...)

nchannels(x, ...)

component_names(x, ...)

ncomponents(x, ...)

nsamples(x, ...)
}
\arguments{
\item{x}{An eeg_lst object.}

\item{...}{Not in use.}
}
\description{
Display information of the eeg_lst object.
\code{*_names()} functions return a vector of names, \code{n*()} return the number of
elements channels or components. Components are only available after running \code{\link[=eeg_ica]{eeg_ica()}}. \code{channel_ica_names()} refers to channels used in the ICA.
}
\seealso{
Other summary functions: \code{\link{count_complete_cases_tbl}},
  \code{\link{eeg_ica_cor_lst}},
  \code{\link{summary.eeg_lst}}
}
\concept{summary functions}
