% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out.R
\name{summary.eeg_lst}
\alias{summary.eeg_lst}
\title{Summary of eeg_lst information.}
\usage{
\method{summary}{eeg_lst}(object, ...)
}
\arguments{
\item{object}{An eeg_lst object.}

\item{...}{additional arguments affecting the summary produced.}
}
\description{
It's also possible to extract the elements by accessing it as a list.
}
\examples{

summary(data_faces_ERPs)

}
\seealso{
Other summary functions: \code{\link{count_complete_cases_tbl}},
  \code{\link{eeg_ica_cor_lst}}, \code{\link{summaries}}
}
\concept{summary functions}
