% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_eeguana}
\alias{theme_eeguana}
\alias{theme_eeguana2}
\alias{default_theme}
\title{Eeguana ggplot themes}
\usage{
theme_eeguana()

theme_eeguana2()

default_theme()
}
\value{
A ggplot theme.
}
\description{
These are complete light themes based on \code{\link[ggplot2:theme_bw]{ggplot2::theme_bw()}} which control all non-data display.
}
\seealso{
Other plotting functions:
\code{\link[=annotate_electrodes]{annotate_electrodes()}},
\code{\link[=annotate_events]{annotate_events()}},
\code{\link[=annotate_head]{annotate_head()}},
\code{\link[=eeg_downsample]{eeg_downsample()}},
\code{\link[=ggplot.eeg_lst]{ggplot.eeg_lst()}},
\code{\link[=plot.eeg_lst]{plot.eeg_lst()}},
\code{\link[=plot_components]{plot_components()}},
\code{\link[=plot_in_layout]{plot_in_layout()}},
\code{\link[=plot_topo]{plot_topo()}}
}
\concept{plotting functions}
