% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_eeguana}
\alias{theme_eeguana}
\alias{theme_eeguana2}
\title{Eeguana ggplot themes}
\usage{
theme_eeguana()

theme_eeguana2()
}
\value{
A ggplot theme.
}
\description{
These are complete light themes based on \code{\link[ggplot2:theme_bw]{ggplot2::theme_bw()}} which control all non-data display.
}
\seealso{
Other plotting functions: 
\code{\link{annotate_events}()},
\code{\link{annotate_head}()},
\code{\link{eeg_downsample}()},
\code{\link{ggplot.eeg_lst}()},
\code{\link{plot.eeg_lst}()},
\code{\link{plot_components}()},
\code{\link{plot_in_layout}()},
\code{\link{plot_topo}()}
}
\concept{plotting functions}
